@@ -1,19 +1,19 @@
 {
   "name": "@spike1292/nativescript",
   "preferGlobal": true,
-  "version": "2.5.0",
+  "version": "2.5.3",
   "author": "Telerik <support@telerik.com>",
   "description": "Command-line interface for building NativeScript projects",
   "bin": {
-    "tns": "./bin/nativescript.js",
-    "nativescript": "./bin/nativescript.js"
+    "tns": "./bin/tns",
+    "nativescript": "./bin/tns"
   },
   "main": "./lib/nativescript-cli.js",
   "scripts": {
-    "test": "node_modules/.bin/istanbul cover node_modules/mocha/bin/_mocha",
+    "test": "node test-scripts/istanbul.js",
     "postinstall": "node postinstall.js",
     "preuninstall": "node preuninstall.js",
-    "mocha": "mocha",
+    "mocha": "node test-scripts/mocha.js",
     "tsc": "tsc",
     "test-watch": "node ./dev/tsc-to-mocha-watch.js"
   },
@@ -27,51 +27,52 @@
     "mobile"
   ],
   "dependencies": {
-    "bluebird": "2.9.34",
     "bplist-parser": "0.1.0",
     "bufferpack": "0.0.6",
-    "bufferutil": "https://github.com/telerik/bufferutil/tarball/v1.0.1.4",
+    "bufferutil": "https://github.com/telerik/bufferutil/tarball/v1.0.1.5",
     "byline": "4.2.1",
     "chalk": "1.1.0",
+    "chokidar": "^1.6.1",
     "cli-table": "https://github.com/telerik/cli-table/tarball/v0.3.1.2",
     "clui": "0.3.1",
     "colors": "1.1.2",
     "email-validator": "1.0.4",
     "esprima": "2.7.0",
-    "ffi": "https://github.com/icenium/node-ffi/tarball/v2.0.0.4",
-    "fibers": "https://github.com/icenium/node-fibers/tarball/v1.0.15.0",
+    "ffi": "https://github.com/icenium/node-ffi/tarball/v2.0.0.5",
+    "fibers": "https://github.com/icenium/node-fibers/tarball/v1.0.15.1",
     "filesize": "3.1.2",
     "gaze": "1.1.0",
     "glob": "^7.0.3",
     "iconv-lite": "0.4.11",
     "inquirer": "0.9.0",
-    "ios-sim-portable": "~1.6.0",
+    "ios-mobileprovision-finder": "1.0.9",
+    "ios-sim-portable": "~1.7.0",
     "lockfile": "1.0.1",
     "lodash": "4.13.1",
-    "log4js": "0.6.26",
+    "log4js": "1.0.1",
     "marked": "0.3.6",
     "marked-terminal": "2.0.0",
     "minimatch": "3.0.2",
     "mkdirp": "0.5.1",
     "moment": "2.10.6",
     "mute-stream": "0.0.5",
-    "node-inspector": "https://github.com/NativeScript/node-inspector/tarball/v0.7.4.2",
-    "node-uuid": "1.4.3",
     "open": "0.0.5",
     "osenv": "0.1.3",
+    "pbxproj-dom": "1.0.8",
     "plist": "1.1.0",
     "plist-merge-patch": "0.0.9",
     "plistlib": "0.2.1",
     "progress-stream": "1.1.1",
     "properties-parser": "0.2.3",
-    "ref": "https://github.com/icenium/ref/tarball/v1.3.2.2",
-    "ref-struct": "https://github.com/telerik/ref-struct/tarball/v1.0.2.4",
+    "ref": "https://github.com/icenium/ref/tarball/v1.3.2.3",
+    "ref-struct": "https://github.com/telerik/ref-struct/tarball/v1.0.2.5",
     "semver": "5.0.1",
-    "shelljs": "0.7.5",
+    "shelljs": "0.7.6",
     "source-map": "0.5.6",
     "tabtab": "https://github.com/Icenium/node-tabtab/tarball/master",
     "temp": "0.8.3",
-    "utf-8-validate": "https://github.com/telerik/utf-8-validate/tarball/v1.0.1.4",
+    "utf-8-validate": "https://github.com/telerik/utf-8-validate/tarball/v1.0.1.5",
+    "uuid": "3.0.1",
     "winreg": "0.0.17",
     "ws": "0.7.1",
     "xcode": "https://github.com/NativeScript/node-xcode/archive/1.4.0.tar.gz",
@@ -87,18 +88,18 @@
     "grunt-contrib-copy": "1.0.0",
     "grunt-contrib-watch": "1.0.0",
     "grunt-shell": "1.3.0",
-    "grunt-ts": "6.0.0-beta.3",
+    "grunt-ts": "6.0.0-beta.6",
     "grunt-tslint": "3.3.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3",
+    "mocha": "3.1.2",
     "mocha-fibers": "https://github.com/NativeScript/mocha-fibers.git",
     "mocha-typescript": "^1.0.4",
     "should": "7.0.2",
     "tslint": "3.15.1",
-    "typescript": "2.0.7"
+    "typescript": "2.1.4"
   },
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=4.2.1 <5.0.0 || >=5.1.0 <8.0.0"
+    "node": ">=6.0.0 <8.0.0"
   }
 }
