@@ -1,7 +1,7 @@
 import * as path from "path";
 import * as minimatch from "minimatch";
 import * as constants from "../constants";
-import Future = require("fibers/future");
+import * as fs from "fs";
 
 export class AppFilesUpdater {
 	constructor(
@@ -38,15 +38,15 @@ export class AppFilesUpdater {
 	}
 
 	protected readDestinationDir(): string[] {
-		if (this.fs.exists(this.appDestinationDirectoryPath).wait()) {
-			return this.fs.readDirectory(this.appDestinationDirectoryPath).wait();
+		if (this.fs.exists(this.appDestinationDirectoryPath)) {
+			return this.fs.readDirectory(this.appDestinationDirectoryPath);
 		} else {
 			return [];
 		}
 	}
 
 	protected deleteDestinationItem(directoryItem: string): void {
-		this.fs.deleteDirectory(path.join(this.appDestinationDirectoryPath, directoryItem)).wait();
+		this.fs.deleteDirectory(path.join(this.appDestinationDirectoryPath, directoryItem));
 	}
 
 	protected readSourceDir(): string[] {
@@ -75,13 +75,19 @@ export class AppFilesUpdater {
 	}
 
 	protected copyAppSourceFiles(sourceFiles: string[]): void {
-		let copyFileFutures = sourceFiles.map(source => {
+		sourceFiles.map(source => {
 			let destinationPath = path.join(this.appDestinationDirectoryPath, path.relative(this.appSourceDirectoryPath, source));
-			if (this.fs.getFsStats(source).wait().isDirectory()) {
+
+			let exists = fs.lstatSync(source);
+			if (exists.isSymbolicLink()) {
+				source = fs.realpathSync(source);
+				exists = fs.lstatSync(source);
+			}
+			if (exists.isDirectory()) {
 				return this.fs.createDirectory(destinationPath);
 			}
+
 			return this.fs.copyFile(source, destinationPath);
 		});
-		Future.wait(copyFileFutures);
 	}
 }
