@@ -1,4 +1,3 @@
 #!/usr/bin/env node
 
-"use strict";
-require("../lib/nativescript-cli.js");
+require("./tns");
