@@ -7,35 +7,33 @@
 		private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		private $config: IConfiguration,
 		private $usbLiveSyncService: ILiveSyncService,
-		protected $options: IOptions) { }
+		protected $platformService: IPlatformService,
+		protected $options: IOptions,
+		protected $platformsData: IPlatformsData) { }
 
 	execute(args: string[]): IFuture<void> {
-
-		if (this.$options.watch) {
-			this.$options.rebuild = false;
+		if (this.$options.start) {
+			return this.debugService.debug();
 		}
 
-		if (!this.$options.rebuild && !this.$options.start) {
-			this.$config.debugLivesync = true;
-			let applicationReloadAction = (deviceAppData: Mobile.IDeviceAppData, localToDevicePaths: Mobile.ILocalToDevicePathData[]): IFuture<void> => {
-				return (() => {
-					let projectData: IProjectData = this.$injector.resolve("projectData");
+		this.$platformService.deployPlatform(this.$devicesService.platform).wait();
+		this.$config.debugLivesync = true;
+		let applicationReloadAction = (deviceAppData: Mobile.IDeviceAppData): IFuture<void> => {
+			return (() => {
+				let projectData: IProjectData = this.$injector.resolve("projectData");
 
-					this.debugService.debugStop().wait();
+				this.debugService.debugStop().wait();
 
-					let applicationId = deviceAppData.appIdentifier;
-					if (deviceAppData.device.isEmulator && deviceAppData.platform.toLowerCase() === this.$devicePlatformsConstants.iOS.toLowerCase()) {
-						applicationId = projectData.projectName;
-					}
-					deviceAppData.device.applicationManager.stopApplication(applicationId).wait();
+				let applicationId = deviceAppData.appIdentifier;
+				if (deviceAppData.device.isEmulator && deviceAppData.platform.toLowerCase() === this.$devicePlatformsConstants.iOS.toLowerCase()) {
+					applicationId = projectData.projectName;
+				}
+				deviceAppData.device.applicationManager.stopApplication(applicationId).wait();
 
-					this.debugService.debug().wait();
-				}).future<void>()();
-			};
-
-			return this.$usbLiveSyncService.liveSync(this.$devicesService.platform, applicationReloadAction);
-		}
-		return this.debugService.debug();
+				this.debugService.debug().wait();
+			}).future<void>()();
+		};
+		return this.$usbLiveSyncService.liveSync(this.$devicesService.platform, applicationReloadAction);
 	}
 
 	allowedParameters: ICommandParameter[] = [];
@@ -69,8 +67,17 @@ export class DebugIOSCommand extends DebugPlatformCommand {
 		$devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		$config: IConfiguration,
 		$usbLiveSyncService: ILiveSyncService,
-		$options: IOptions) {
-		super($iOSDebugService, $devicesService, $injector, $logger, $childProcess, $devicePlatformsConstants, $config, $usbLiveSyncService, $options);
+		$platformService: IPlatformService,
+		$options: IOptions,
+		$platformsData: IPlatformsData) {
+
+		super($iOSDebugService, $devicesService, $injector, $logger, $childProcess, $devicePlatformsConstants, $config, $usbLiveSyncService, $platformService, $options, $platformsData);
+	}
+
+	canExecute(args: string[]): IFuture<boolean> {
+		return (() => {
+			return super.canExecute(args).wait() && this.$platformService.validateOptions(this.$platformsData.availablePlatforms.iOS).wait();
+		}).future<boolean>()();
 	}
 }
 $injector.registerCommand("debug|ios", DebugIOSCommand);
@@ -84,8 +91,17 @@ export class DebugAndroidCommand extends DebugPlatformCommand {
 		$devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		$config: IConfiguration,
 		$usbLiveSyncService: ILiveSyncService,
-		$options: IOptions) {
-		super($androidDebugService, $devicesService, $injector, $logger, $childProcess, $devicePlatformsConstants, $config, $usbLiveSyncService, $options);
+		$platformService: IPlatformService,
+		$options: IOptions,
+		$platformsData: IPlatformsData) {
+
+		super($androidDebugService, $devicesService, $injector, $logger, $childProcess, $devicePlatformsConstants, $config, $usbLiveSyncService, $platformService, $options, $platformsData);
+	}
+
+	canExecute(args: string[]): IFuture<boolean> {
+		return (() => {
+			return super.canExecute(args).wait() && this.$platformService.validateOptions(this.$platformsData.availablePlatforms.Android).wait();
+		}).future<boolean>()();
 	}
 }
 $injector.registerCommand("debug|android", DebugAndroidCommand);
