@@ -6,10 +6,10 @@ export class ListPlatformsCommand implements ICommand {
 
 	execute(args: string[]): IFuture<void> {
 		return (() => {
-			let installedPlatforms = this.$platformService.getInstalledPlatforms().wait();
+			let installedPlatforms = this.$platformService.getInstalledPlatforms();
 
 			if(installedPlatforms.length > 0) {
-				let preparedPlatforms = this.$platformService.getPreparedPlatforms().wait();
+				let preparedPlatforms = this.$platformService.getPreparedPlatforms();
 				if(preparedPlatforms.length > 0) {
 					this.$logger.out("The project is prepared for: ", helpers.formatListOfNames(preparedPlatforms, "and"));
 				} else {
@@ -17,7 +17,7 @@ export class ListPlatformsCommand implements ICommand {
 				}
 				this.$logger.out("Installed platforms: ", helpers.formatListOfNames(installedPlatforms, "and"));
 			} else {
-				let formattedPlatformsList = helpers.formatListOfNames(this.$platformService.getAvailablePlatforms().wait(), "and");
+				let formattedPlatformsList = helpers.formatListOfNames(this.$platformService.getAvailablePlatforms(), "and");
 				this.$logger.out("Available platforms for this OS: ", formattedPlatformsList);
 				this.$logger.out("No installed platforms found. Use $ tns platform add");
 			}
