@@ -8,6 +8,7 @@ export class UpdateCommand implements ICommand {
 		private $platformsData: IPlatformsData,
 		private $pluginsService: IPluginsService,
 		private $projectDataService: IProjectDataService,
+		private $fs: IFileSystem,
 		private $logger: ILogger,
 		private $options: IOptions,
 		private $errors: IErrors) { }
@@ -22,7 +23,10 @@ export class UpdateCommand implements ICommand {
 				shelljs.mkdir(tmpDir);
 				shelljs.cp(path.join(this.$projectData.projectDir, "package.json"), tmpDir);
 				for (let folder of folders) {
-					shelljs.cp("-rf", path.join(this.$projectData.projectDir, folder), tmpDir);
+					let folderToCopy = path.join(this.$projectData.projectDir, folder);
+					if (this.$fs.exists(folderToCopy)) {
+						shelljs.cp("-rf", folderToCopy, tmpDir);
+					}
 				}
 			} catch(error) {
 				this.$logger.error("Could not backup project folders!");
@@ -30,12 +34,15 @@ export class UpdateCommand implements ICommand {
 			}
 
 			try {
-				this.executeCore(args, folders);
+				this.executeCore(args, folders).wait();
 			} catch (error) {
 				shelljs.cp("-f", path.join(tmpDir, "package.json"), this.$projectData.projectDir);
 				for (let folder of folders) {
 					shelljs.rm("-rf", path.join(this.$projectData.projectDir, folder));
-					shelljs.cp("-fr", path.join(tmpDir, folder), this.$projectData.projectDir);
+					let folderToCopy = path.join(tmpDir, folder);
+					if (this.$fs.exists(folderToCopy)) {
+						shelljs.cp("-fr", folderToCopy, this.$projectData.projectDir);
+					}
 				}
 				this.$logger.error("Could not update the project!");
 			} finally {
@@ -51,40 +58,42 @@ export class UpdateCommand implements ICommand {
 		}).future<boolean>()();
 	}
 
-	private executeCore(args: string[], folders: string[]) {
-		let platforms = this.$platformService.getInstalledPlatforms().wait();
-		let availablePlatforms = this.$platformService.getAvailablePlatforms().wait();
-		let packagePlatforms: string[] = [];
+	private executeCore(args: string[], folders: string[]): IFuture<void> {
+		return (() => {
+			let platforms = this.$platformService.getInstalledPlatforms();
+			let availablePlatforms = this.$platformService.getAvailablePlatforms();
+			let packagePlatforms: string[] = [];
 
-		this.$projectDataService.initialize(this.$projectData.projectDir);
-		for (let platform of availablePlatforms) {
-			let platformData = this.$platformsData.getPlatformData(platform);
-			let platformVersion = this.$projectDataService.getValue(platformData.frameworkPackageName).wait();
-			if (platformVersion) {
-				packagePlatforms.push(platform);
-				this.$projectDataService.removeProperty(platformData.frameworkPackageName).wait();
+			this.$projectDataService.initialize(this.$projectData.projectDir);
+			for (let platform of availablePlatforms) {
+				let platformData = this.$platformsData.getPlatformData(platform);
+				let platformVersion = this.$projectDataService.getValue(platformData.frameworkPackageName);
+				if (platformVersion) {
+					packagePlatforms.push(platform);
+					this.$projectDataService.removeProperty(platformData.frameworkPackageName);
+				}
 			}
-		}
 
-		this.$platformService.removePlatforms(platforms).wait();
-		this.$pluginsService.remove("tns-core-modules").wait();
-		this.$pluginsService.remove("tns-core-modules-widgets").wait();
+			this.$platformService.removePlatforms(platforms);
+			this.$pluginsService.remove("tns-core-modules").wait();
+			this.$pluginsService.remove("tns-core-modules-widgets").wait();
 
-		for (let folder of folders) {
-			shelljs.rm("-fr", folder);
-		}
-
-		platforms = platforms.concat(packagePlatforms);
-		if (args.length === 1) {
-			for (let platform of platforms) {
-				this.$platformService.addPlatforms([ platform+"@"+args[0] ]).wait();
+			for (let folder of folders) {
+				shelljs.rm("-fr", folder);
 			}
-			this.$pluginsService.add("tns-core-modules@" + args[0]).wait();
-		} else {
-			this.$platformService.addPlatforms(platforms).wait();
-			this.$pluginsService.add("tns-core-modules").wait();
-		}
-		this.$pluginsService.ensureAllDependenciesAreInstalled().wait();
+
+			platforms = platforms.concat(packagePlatforms);
+			if (args.length === 1) {
+				for (let platform of platforms) {
+					this.$platformService.addPlatforms([ platform+"@"+args[0] ]).wait();
+				}
+				this.$pluginsService.add("tns-core-modules@" + args[0]).wait();
+			} else {
+				this.$platformService.addPlatforms(platforms).wait();
+				this.$pluginsService.add("tns-core-modules").wait();
+			}
+			this.$pluginsService.ensureAllDependenciesAreInstalled().wait();
+		}).future<void>()();
 	}
 
 	allowedParameters: ICommandParameter[] = [];
