@@ -1,16 +1,15 @@
 import * as constants from './constants';
-import Future = require("fibers/future");
 
 export class DynamicHelpProvider implements IDynamicHelpProvider {
-	public isProjectType(args: string[]): IFuture<boolean> {
-		return Future.fromResult(true);
+	public isProjectType(args: string[]): boolean {
+		return true;
 	}
 
-	public getLocalVariables(options: { isHtml: boolean }): IFuture<IDictionary<any>> {
+	public getLocalVariables(options: { isHtml: boolean }): IDictionary<any> {
 		let localVariables: IDictionary<any> = {
 			constants: constants
 		};
-		return Future.fromResult(localVariables);
+		return localVariables;
 	}
 }
 $injector.register("dynamicHelpProvider", DynamicHelpProvider);
