@@ -1,12 +1,9 @@
-let node = require("../package.json").engines.node;
-// this call must be first to avoid requiring c++ dependencies
-require("./common/verify-node-version").verifyNodeVersion(node, "NativeScript", "2.5.0");
-
 require("./bootstrap");
 import * as fiber from "fibers";
 import Future = require("fibers/future");
 import * as shelljs from "shelljs";
 shelljs.config.silent = true;
+shelljs.config.fatal = true;
 import {installUncaughtExceptionListener} from "./common/errors";
 installUncaughtExceptionListener(process.exit);
 
