@@ -1,14 +1,11 @@
 import * as path from "path";
 import * as semver from "semver";
 import * as constants from "./constants";
-import {sleep} from "../lib/common/helpers";
 
 export class NpmInstallationManager implements INpmInstallationManager {
-	private static NPM_LOAD_FAILED = "Failed to retrieve data from npm. Please try again a little bit later.";
-
 	constructor(private $npm: INodePackageManager,
+		private $childProcess: IChildProcess,
 		private $logger: ILogger,
-		private $lockfile: ILockFile,
 		private $errors: IErrors,
 		private $options: IOptions,
 		private $fs: IFileSystem,
@@ -16,7 +13,7 @@ export class NpmInstallationManager implements INpmInstallationManager {
 	}
 
 	public getLatestVersion(packageName: string): IFuture<string> {
-		return(() => {
+		return (() => {
 			return this.getVersion(packageName, constants.PackageVersion.LATEST).wait();
 		}).future<string>()();
 	}
@@ -32,10 +29,10 @@ export class NpmInstallationManager implements INpmInstallationManager {
 
 			let cliVersionRange = `~${this.$staticConfig.version}`;
 			let latestVersion = this.getLatestVersion(packageName).wait();
-			if(semver.satisfies(latestVersion, cliVersionRange)) {
+			if (semver.satisfies(latestVersion, cliVersionRange)) {
 				return latestVersion;
 			}
-			let data = this.$npm.view(packageName, {json: true, "versions": true}).wait();
+			let data = this.$npm.view(packageName, { json: true, "versions": true }).wait();
 
 			return semver.maxSatisfying(data, cliVersionRange) || latestVersion;
 		}).future<string>()();
@@ -44,13 +41,6 @@ export class NpmInstallationManager implements INpmInstallationManager {
 	public install(packageName: string, projectDir: string, opts?: INpmInstallOptions): IFuture<any> {
 		return (() => {
 
-			// TODO: plamen5kov: figure a way to remove this
-			while(this.$lockfile.check().wait()) {
-				sleep(10);
-			}
-
-			this.$lockfile.lock().wait();
-
 			try {
 				let packageToInstall = this.$options.frameworkPath || packageName;
 				let pathToSave = projectDir;
@@ -58,34 +48,64 @@ export class NpmInstallationManager implements INpmInstallationManager {
 				let dependencyType = (opts && opts.dependencyType) || null;
 
 				return this.installCore(packageToInstall, pathToSave, version, dependencyType).wait();
-			} catch(error) {
+			} catch (error) {
 				this.$logger.debug(error);
-				this.$errors.fail("%s. Error: %s", NpmInstallationManager.NPM_LOAD_FAILED, error);
-			} finally {
-				this.$lockfile.unlock().wait();
+
+				throw new Error(error);
 			}
 
 		}).future<string>()();
 	}
 
+	public getInspectorFromCache(inspectorNpmPackageName: string, projectDir: string): IFuture<string> {
+		return (() => {
+			let inspectorPath = path.join(projectDir, "node_modules", inspectorNpmPackageName);
+
+			// local installation takes precedence over cache
+			if (!this.inspectorAlreadyInstalled(inspectorPath)) {
+				let cachepath = this.$childProcess.exec("npm get cache").wait().trim();
+				let version = this.getLatestCompatibleVersion(inspectorNpmPackageName).wait();
+				let pathToPackageInCache = path.join(cachepath, inspectorNpmPackageName, version);
+				let pathToUnzippedInspector = path.join(pathToPackageInCache, "package");
+
+				if (!this.$fs.exists(pathToPackageInCache)) {
+					this.$childProcess.exec(`npm cache add ${inspectorNpmPackageName}@${version}`).wait();
+					let inspectorTgzPathInCache = path.join(pathToPackageInCache, "package.tgz");
+					this.$childProcess.exec(`tar -xf ${inspectorTgzPathInCache} -C ${pathToPackageInCache}`).wait();
+					this.$childProcess.exec(`npm install --prefix ${pathToUnzippedInspector}`).wait();
+				}
+				this.$logger.out("Using inspector from cache.");
+				return pathToUnzippedInspector;
+			}
+			return inspectorPath;
+		}).future<string>()();
+	}
+
+	private inspectorAlreadyInstalled(pathToInspector: string): Boolean {
+		if (this.$fs.exists(pathToInspector)) {
+			return true;
+		}
+		return false;
+	}
+
 	private installCore(packageName: string, pathToSave: string, version: string, dependencyType: string): IFuture<string> {
 		return (() => {
-			const possiblePackageName= path.resolve(packageName);
-			if(this.$fs.exists(possiblePackageName).wait()) {
+			const possiblePackageName = path.resolve(packageName);
+			if (this.$fs.exists(possiblePackageName)) {
 				packageName = possiblePackageName;
 			}
-			if(packageName.indexOf(".tgz") >= 0) {
+			if (packageName.indexOf(".tgz") >= 0) {
 				version = null;
 			}
 			// check if the packageName is url or local file and if it is, let npm install deal with the version
-			if(this.isURL(packageName) || this.$fs.exists(packageName).wait()) {
+			if (this.isURL(packageName) || this.$fs.exists(packageName)) {
 				version = null;
 			} else {
 				version = version || this.getLatestCompatibleVersion(packageName).wait();
 			}
 
 			let installedModuleNames = this.npmInstall(packageName, pathToSave, version, dependencyType).wait();
-			let installedPackageName =  installedModuleNames[0];
+			let installedPackageName = installedModuleNames[0];
 
 			let pathToInstalledPackage = path.join(pathToSave, "node_modules", installedPackageName);
 			return pathToInstalledPackage;
@@ -99,18 +119,18 @@ export class NpmInstallationManager implements INpmInstallationManager {
 	}
 
 	private npmInstall(packageName: string, pathToSave: string, version: string, dependencyType: string): IFuture<any> {
-		return(() => {
+		return (() => {
 			this.$logger.out("Installing ", packageName);
 
 			packageName = packageName + (version ? `@${version}` : "");
 
-			let npmOptions: any = {silent: true};
+			let npmOptions: any = { silent: true };
 
-			if(dependencyType) {
+			if (dependencyType) {
 				npmOptions[dependencyType] = true;
 			}
 
-			return this.$npm.install(packageName , pathToSave, npmOptions).wait();
+			return this.$npm.install(packageName, pathToSave, npmOptions).wait();
 		}).future<any>()();
 	}
 
@@ -120,7 +140,7 @@ export class NpmInstallationManager implements INpmInstallationManager {
 	 */
 	private getVersion(packageName: string, version: string): IFuture<string> {
 		return (() => {
-			let data:any = this.$npm.view(packageName, {json: true, "dist-tags": true}).wait();
+			let data: any = this.$npm.view(packageName, { json: true, "dist-tags": true }).wait();
 			this.$logger.trace("Using version %s. ", data[version]);
 
 			return data[version];
