@@ -17,6 +17,7 @@ export class Options extends commonOptionsLibPath.OptionsBase {
 			copyFrom: { type: OptionType.String },
 			linkTo: { type: OptionType.String },
 			forDevice: { type: OptionType.Boolean },
+			provision: { type: OptionType.Object },
 			client: { type: OptionType.Boolean, default: true },
 			production: { type: OptionType.Boolean },
 			debugTransport: { type: OptionType.Boolean },
@@ -38,10 +39,11 @@ export class Options extends commonOptionsLibPath.OptionsBase {
 			bundle: { type: OptionType.Boolean },
 			all: { type: OptionType.Boolean },
 			teamId: { type: OptionType.String },
-			rebuild: { type: OptionType.Boolean, default: true },
-			syncAllFiles: { type: OptionType.Boolean },
+			syncAllFiles: { type: OptionType.Boolean, default: false },
 			liveEdit: { type: OptionType.Boolean },
-			chrome: { type: OptionType.Boolean }
+			chrome: { type: OptionType.Boolean },
+			clean: { type: OptionType.Boolean },
+			watch: { type: OptionType.Boolean, default: true }
 		},
 			path.join($hostInfo.isWindows ? process.env.AppData : path.join(osenv.home(), ".local/share"), ".nativescript-cli"),
 			$errors, $staticConfig);
