@@ -3,9 +3,7 @@ import * as temp from "temp";
 
 export class LiveSyncProvider implements ILiveSyncProvider {
 	constructor(private $androidLiveSyncServiceLocator: {factory: Function},
-		private $androidPlatformLiveSyncServiceLocator: {factory: Function},
 		private $iosLiveSyncServiceLocator: {factory: Function},
-		private $iosPlatformLiveSyncServiceLocator: {factory: Function},
 		private $platformService: IPlatformService,
 		private $platformsData: IPlatformsData,
 		private $logger: ILogger,
@@ -35,35 +33,15 @@ export class LiveSyncProvider implements ILiveSyncProvider {
 		};
 	}
 
-	private platformSpecificLiveSyncServicesCache: IDictionary<any> = {};
-	public get platformSpecificLiveSyncServices(): IDictionary<any> {
-		return {
-			android: (_liveSyncData: ILiveSyncData, $injector: IInjector) => {
-				if(!this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.Android]) {
-					this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.Android] = $injector.resolve(this.$androidPlatformLiveSyncServiceLocator.factory, { _liveSyncData: _liveSyncData });
-				}
-
-				return this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.Android];
-			},
-			ios: (_liveSyncData: ILiveSyncData, $injector: IInjector) => {
-				if(!this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.iOS]) {
-					this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.iOS] = $injector.resolve(this.$iosPlatformLiveSyncServiceLocator.factory, { _liveSyncData: _liveSyncData });
-				}
-
-				return this.platformSpecificLiveSyncServicesCache[this.$devicePlatformsConstants.iOS];
-			}
-		};
-	}
-
 	public buildForDevice(device: Mobile.IDevice): IFuture<string> {
 		return (() => {
-			this.$platformService.buildForDeploy(device.deviceInfo.platform, {buildForDevice: !device.isEmulator}).wait();
+			this.$platformService.buildPlatform(device.deviceInfo.platform, {buildForDevice: !device.isEmulator}).wait();
 			let platformData = this.$platformsData.getPlatformData(device.deviceInfo.platform);
 			if (device.isEmulator) {
-				return this.$platformService.getLatestApplicationPackageForEmulator(platformData).wait().packageName;
+				return this.$platformService.getLatestApplicationPackageForEmulator(platformData).packageName;
 			}
 
-			return this.$platformService.getLatestApplicationPackageForDevice(platformData).wait().packageName;
+			return this.$platformService.getLatestApplicationPackageForDevice(platformData).packageName;
 		}).future<string>()();
 	}
 
@@ -91,7 +69,8 @@ export class LiveSyncProvider implements ILiveSyncProvider {
 				if (this.$options.syncAllFiles) {
 					this.$childProcess.spawnFromEvent("zip", [ "-r", "-0", tempZip, "app" ], "close", { cwd: path.dirname(projectFilesPath) }).wait();
 				} else {
-					this.$childProcess.spawnFromEvent("zip", [ "-r", "-0", tempZip, "app", "-x", "app/tns_modules/*" ], "close", { cwd: path.dirname(projectFilesPath) }).wait();
+				 	this.$logger.info("Skipping node_modules folder! Use the syncAllFiles option to sync files from this folder.");
+				 	this.$childProcess.spawnFromEvent("zip", [ "-r", "-0", tempZip, "app", "-x", "app/tns_modules/*" ], "close", { cwd: path.dirname(projectFilesPath) }).wait();
 				}
 
 				deviceAppData.device.fileSystem.transferFiles(deviceAppData, [{
