@@ -1,12 +1,10 @@
-import * as path from "path";
 import * as net from "net";
+import * as os from "os";
 import Future = require("fibers/future");
 import { sleep } from "../common/helpers";
 import {ChildProcess} from "child_process";
 
 class AndroidDebugService implements IDebugService {
-	private static DEFAULT_NODE_INSPECTOR_URL = "http://127.0.0.1:8080/debug";
-
 	private _device: Mobile.IAndroidDevice = null;
 	private _debuggerClientProcess: ChildProcess;
 
@@ -42,7 +40,6 @@ class AndroidDebugService implements IDebugService {
 
 	private debugOnEmulator(): IFuture<void> {
 		return (() => {
-			this.$platformService.deployOnEmulator(this.platform).wait();
 			// Assure we've detected the emulator as device
 			// For example in case deployOnEmulator had stated new emulator instance
 			// we need some time to detect it. Let's force detection.
@@ -78,8 +75,10 @@ class AndroidDebugService implements IDebugService {
 			let port = -1;
 			let forwardsResult = this.device.adb.executeCommand(["forward", "--list"]).wait();
 
+			let unixSocketName = `${packageName}-inspectorServer`;
+
 			//matches 123a188909e6czzc tcp:40001 localabstract:org.nativescript.testUnixSockets-debug
-			let regexp = new RegExp(`(?:${deviceId} tcp:)([\\d]+)(?= localabstract:${packageName}-debug)`, "g");
+			let regexp = new RegExp(`(?:${deviceId} tcp:)([\\d]+)(?= localabstract:${unixSocketName})`, "g");
 			let match = regexp.exec(forwardsResult);
 			if (match) {
 				port = parseInt(match[1]);
@@ -92,7 +91,7 @@ class AndroidDebugService implements IDebugService {
 				}
 				port = candidatePort;
 
-				this.unixSocketForward(port, packageName + "-debug").wait();
+				this.unixSocketForward(port, `${unixSocketName}`).wait();
 			}
 
 			return port;
@@ -107,16 +106,13 @@ class AndroidDebugService implements IDebugService {
 		return (() => {
 			let packageFile = "";
 
-			if (!this.$options.start && !this.$options.emulator && !this.$options.getPort) {
+			if (!this.$options.start && !this.$options.emulator) {
 				let cachedDeviceOption = this.$options.forDevice;
 				this.$options.forDevice = true;
-				if (this.$options.rebuild) {
-					this.$platformService.prepareAndBuild(this.platform).wait();
-				}
 				this.$options.forDevice = !!cachedDeviceOption;
 
 				let platformData = this.$platformsData.getPlatformData(this.platform);
-				packageFile = this.$platformService.getLatestApplicationPackageForDevice(platformData).wait().packageName;
+				packageFile = this.$platformService.getLatestApplicationPackageForDevice(platformData).packageName;
 				this.$logger.out("Using ", packageFile);
 			}
 
@@ -131,9 +127,9 @@ class AndroidDebugService implements IDebugService {
 		return (() => {
 			this.device = device;
 
-			if (this.$options.getPort) {
-				this.printDebugPort(device.deviceInfo.identifier, packageName).wait();
-			} else if (this.$options.start) {
+			this.printDebugPort(device.deviceInfo.identifier, packageName).wait();
+
+			if (this.$options.start) {
 				this.attachDebugger(device.deviceInfo.identifier, packageName).wait();
 			} else if (this.$options.stop) {
 				this.detachDebugger(packageName).wait();
@@ -159,7 +155,6 @@ class AndroidDebugService implements IDebugService {
 			if (this.$options.client) {
 				let port = this.getForwardedLocalDebugPortForPackageName(deviceId, packageName).wait();
 				this.startDebuggerClient(port).wait();
-				this.openDebuggerClient(AndroidDebugService.DEFAULT_NODE_INSPECTOR_URL + "?port=" + port);
 			}
 		}).future<void>()();
 	}
@@ -238,11 +233,7 @@ class AndroidDebugService implements IDebugService {
 
 	private startDebuggerClient(port: Number): IFuture<void> {
 		return (() => {
-			let nodeInspectorModuleFilePath = require.resolve("node-inspector");
-			let nodeInspectorModuleDir = path.dirname(nodeInspectorModuleFilePath);
-			let nodeInspectorFullPath = path.join(nodeInspectorModuleDir, "bin", "inspector");
-			this._debuggerClientProcess = this.$childProcess.spawn(process.argv[0], [nodeInspectorFullPath, "--debug-port", port.toString()], { stdio: "ignore", detached: true });
-			this.$processService.attachToProcessExitSignals(this, this.debugStop);
+			this.$logger.info(`To start debugging, open the following URL in Chrome:${os.EOL}chrome-devtools://devtools/bundled/inspector.html?experiments=true&v8only=true&ws=localhost:${port}${os.EOL}`.cyan);
 		}).future<void>()();
 	}
 
@@ -253,20 +244,5 @@ class AndroidDebugService implements IDebugService {
 		}
 	}
 
-	private openDebuggerClient(url: string): void {
-		let defaultDebugUI = "chrome";
-		if (this.$hostInfo.isDarwin) {
-			defaultDebugUI = "Google Chrome";
-		}
-		if (this.$hostInfo.isLinux) {
-			defaultDebugUI = "google-chrome";
-		}
-
-		let debugUI = this.$config.ANDROID_DEBUG_UI || defaultDebugUI;
-		let child = this.$opener.open(url, debugUI);
-		if (!child) {
-			this.$errors.failWithoutHelp(`Unable to open ${debugUI}.`);
-		}
-	}
 }
 $injector.register("androidDebugService", AndroidDebugService);
