@@ -8,9 +8,11 @@ import * as helpers from "../common/helpers";
 import * as projectServiceBaseLib from "./platform-project-service-base";
 import Future = require("fibers/future");
 import { PlistSession } from "plist-merge-patch";
-import {EOL} from "os";
+import { EOL } from "os";
 import * as temp from "temp";
 import * as plist from "plist";
+import { Xcode } from "pbxproj-dom/xcode";
+import { IOSProvisionService } from "./ios-provision-service";
 
 export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServiceBase implements IPlatformProjectService {
 	private static XCODE_PROJECT_EXT_NAME = ".xcodeproj";
@@ -41,7 +43,8 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 		private $devicesService: Mobile.IDevicesService,
 		private $mobileHelper: Mobile.IMobileHelper,
 		private $pluginVariablesService: IPluginVariablesService,
-		private $xcprojService: IXcprojService) {
+		private $xcprojService: IXcprojService,
+		private $iOSProvisionService: IOSProvisionService) {
 		super($fs, $projectData, $projectDataService);
 	}
 
@@ -74,16 +77,25 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 		};
 	}
 
-	public getAppResourcesDestinationDirectoryPath(): IFuture<string> {
+	public validateOptions(): IFuture<boolean> {
 		return (() => {
-			let frameworkVersion = this.getFrameworkVersion(this.platformData.frameworkPackageName).wait();
-
-			if (semver.lt(frameworkVersion, "1.3.0")) {
-				return path.join(this.platformData.projectRoot, this.$projectData.projectName, "Resources", "icons");
+			if (this.$options.provision === true) {
+				this.$iOSProvisionService.list().wait();
+				this.$errors.failWithoutHelp("Please provide provisioning profile uuid or name with the --provision option.");
+				return false;
 			}
+			return true;
+		}).future<boolean>()();
+	}
 
-			return path.join(this.platformData.projectRoot, this.$projectData.projectName, "Resources");
-		}).future<string>()();
+	public getAppResourcesDestinationDirectoryPath(): string {
+		let frameworkVersion = this.getFrameworkVersion(this.platformData.frameworkPackageName);
+
+		if (semver.lt(frameworkVersion, "1.3.0")) {
+			return path.join(this.platformData.projectRoot, this.$projectData.projectName, "Resources", "icons");
+		}
+
+		return path.join(this.platformData.projectRoot, this.$projectData.projectName, "Resources");
 	}
 
 	public validate(): IFuture<void> {
@@ -102,12 +114,13 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 		}).future<void>()();
 	}
 
+	// TODO: Remove IFuture, reason: readDirectory - unable until androidProjectService has async operations.
 	public createProject(frameworkDir: string, frameworkVersion: string, pathToTemplate?: string): IFuture<void> {
 		return (() => {
-			this.$fs.ensureDirectoryExists(path.join(this.platformData.projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER)).wait();
+			this.$fs.ensureDirectoryExists(path.join(this.platformData.projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER));
 			if (pathToTemplate) {
 				// Copy everything except the template from the runtime
-				this.$fs.readDirectory(frameworkDir).wait()
+				this.$fs.readDirectory(frameworkDir)
 					.filter(dirName => dirName.indexOf(IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER) === -1)
 					.forEach(dirName => shell.cp("-R", path.join(frameworkDir, dirName), this.platformData.projectRoot));
 				shell.cp("-rf", path.join(pathToTemplate, "*"), this.platformData.projectRoot);
@@ -116,53 +129,46 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			}
 		}).future<void>()();
 	}
-
+	//TODO: plamen5kov: revisit this method, might have unnecessary/obsolete logic
 	public interpolateData(): IFuture<void> {
 		return (() => {
-			let infoPlistFilePath = path.join(this.platformData.projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER, `${IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER}-Info.plist`);
-			this.interpolateConfigurationFile(infoPlistFilePath).wait();
-
 			let projectRootFilePath = path.join(this.platformData.projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER);
 			// Starting with NativeScript for iOS 1.6.0, the project Info.plist file resides not in the platform project,
 			// but in the hello-world app template as a platform specific resource.
-			if (this.$fs.exists(path.join(projectRootFilePath, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + "-Info.plist")).wait()) {
-				this.replaceFileName("-Info.plist", projectRootFilePath).wait();
+			if (this.$fs.exists(path.join(projectRootFilePath, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + "-Info.plist"))) {
+				this.replaceFileName("-Info.plist", projectRootFilePath);
 			}
-			this.replaceFileName("-Prefix.pch", projectRootFilePath).wait();
+			this.replaceFileName("-Prefix.pch", projectRootFilePath);
 
 			let xcschemeDirPath = path.join(this.platformData.projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + IOSProjectService.XCODE_PROJECT_EXT_NAME, "xcshareddata/xcschemes");
 			let xcschemeFilePath = path.join(xcschemeDirPath, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + IOSProjectService.XCODE_SCHEME_EXT_NAME);
 
-			if (this.$fs.exists(xcschemeFilePath).wait()) {
+			if (this.$fs.exists(xcschemeFilePath)) {
 				this.$logger.debug("Found shared scheme at xcschemeFilePath, renaming to match project name.");
 				this.$logger.debug("Checkpoint 0");
-				this.replaceFileContent(xcschemeFilePath).wait();
+				this.replaceFileContent(xcschemeFilePath);
 				this.$logger.debug("Checkpoint 1");
-				this.replaceFileName(IOSProjectService.XCODE_SCHEME_EXT_NAME, xcschemeDirPath).wait();
+				this.replaceFileName(IOSProjectService.XCODE_SCHEME_EXT_NAME, xcschemeDirPath);
 				this.$logger.debug("Checkpoint 2");
 			} else {
 				this.$logger.debug("Copying xcscheme from template not found at " + xcschemeFilePath);
 			}
 
-			this.replaceFileName(IOSProjectService.XCODE_PROJECT_EXT_NAME, this.platformData.projectRoot).wait();
+			this.replaceFileName(IOSProjectService.XCODE_PROJECT_EXT_NAME, this.platformData.projectRoot);
 
 			let pbxprojFilePath = this.pbxProjPath;
-			this.replaceFileContent(pbxprojFilePath).wait();
+			this.replaceFileContent(pbxprojFilePath);
 
 		}).future<void>()();
 	}
 
 	public interpolateConfigurationFile(configurationFilePath?: string): IFuture<void> {
-		return (() => {
-			shell.sed('-i', "__CFBUNDLEIDENTIFIER__", this.$projectData.projectId, configurationFilePath || this.platformData.configurationFilePath);
-		}).future<void>()();
+		return Future.fromResult();
 	}
 
-	public afterCreateProject(projectRoot: string): IFuture<void> {
-		return (() => {
-			this.$fs.rename(path.join(projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER),
-				path.join(projectRoot, this.$projectData.projectName)).wait();
-		}).future<void>()();
+	public afterCreateProject(projectRoot: string): void {
+		this.$fs.rename(path.join(projectRoot, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER),
+			path.join(projectRoot, this.$projectData.projectName));
 	}
 
 	/**
@@ -214,7 +220,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			// Save the options...
 			temp.track();
 			let exportOptionsPlist = temp.path({ prefix: "export-", suffix: ".plist" });
-			this.$fs.writeFile(exportOptionsPlist, plistTemplate).wait();
+			this.$fs.writeFile(exportOptionsPlist, plistTemplate);
 
 			let args = ["-exportArchive",
 				"-archivePath", archivePath,
@@ -229,7 +235,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 
 	private xcbuildProjectArgs(projectRoot: string, product?: "scheme" | "target"): string[] {
 		let xcworkspacePath = path.join(projectRoot, this.$projectData.projectName + ".xcworkspace");
-		if (this.$fs.exists(xcworkspacePath).wait()) {
+		if (this.$fs.exists(xcworkspacePath)) {
 			return ["-workspace", xcworkspacePath, product ? "-" + product : "-scheme", this.$projectData.projectName];
 		} else {
 			let xcodeprojPath = path.join(projectRoot, this.$projectData.projectName + ".xcodeproj");
@@ -244,46 +250,70 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 				"build",
 				'SHARED_PRECOMPS_DIR=' + path.join(projectRoot, 'build', 'sharedpch')
 			];
-
 			basicArgs = basicArgs.concat(this.xcbuildProjectArgs(projectRoot));
 
 			// Starting from tns-ios 1.4 the xcconfig file is referenced in the project template
-			let frameworkVersion = this.getFrameworkVersion(this.platformData.frameworkPackageName).wait();
+			let frameworkVersion = this.getFrameworkVersion(this.platformData.frameworkPackageName);
 			if (semver.lt(frameworkVersion, "1.4.0")) {
 				basicArgs.push("-xcconfig", path.join(projectRoot, this.$projectData.projectName, "build.xcconfig"));
 			}
 
-			let args: string[] = [];
+			// if (this.$logger.getLevel() === "INFO") {
+			// 	let xcodeBuildVersion = this.getXcodeVersion();
+			// 	if (helpers.versionCompare(xcodeBuildVersion, "8.0") >= 0) {
+			// 		basicArgs.push("-quiet");
+			// 	}
+			// }
+
 			let buildForDevice = this.$options.forDevice || (buildConfig && buildConfig.buildForDevice);
 			if (buildForDevice) {
-				let defaultArchitectures = [
-					'ARCHS=armv7 arm64',
-					'VALID_ARCHS=armv7 arm64'
-				];
-
-				args = basicArgs.concat([
-					"-sdk", "iphoneos",
-					"CONFIGURATION_BUILD_DIR=" + path.join(projectRoot, "build", "device")
-				]);
-
-				args = args.concat((buildConfig && buildConfig.architectures) || defaultArchitectures);
-
-				let xcodeBuildVersion = this.getXcodeVersion();
-				if (helpers.versionCompare(xcodeBuildVersion, "8.0") >= 0) {
-					let teamId = this.getDevelopmentTeam();
-					if (teamId) {
-						args = args.concat("DEVELOPMENT_TEAM=" + teamId);
-					}
-				}
+				this.buildForDevice(projectRoot, basicArgs, buildConfig).wait();
 			} else {
-				args = basicArgs.concat([
-					"-sdk", "iphonesimulator",
-					"ARCHS=i386 x86_64",
-					"VALID_ARCHS=i386 x86_64",
-					"ONLY_ACTIVE_ARCH=NO",
-					"CONFIGURATION_BUILD_DIR=" + path.join(projectRoot, "build", "emulator"),
-					"CODE_SIGN_IDENTITY="
-				]);
+				this.buildForSimulator(projectRoot, basicArgs, buildConfig).wait();
+			}
+		}).future<void>()();
+	}
+
+	private buildForDevice(projectRoot: string, args: string[], buildConfig?: IiOSBuildConfig): IFuture<void> {
+		return (() => {
+			let defaultArchitectures = [
+				'ARCHS=armv7 arm64',
+				'VALID_ARCHS=armv7 arm64'
+			];
+
+			// build only for device specific architecture
+			if (!this.$options.release && !(buildConfig && buildConfig.architectures)) {
+				this.$devicesService.initialize({ platform: this.$devicePlatformsConstants.iOS.toLowerCase(), deviceId: this.$options.device }).wait();
+				let instances = this.$devicesService.getDeviceInstances();
+				let devicesArchitectures = _(instances)
+					.filter(d => this.$mobileHelper.isiOSPlatform(d.deviceInfo.platform) && d.deviceInfo.activeArchitecture)
+					.map(d => d.deviceInfo.activeArchitecture)
+					.uniq()
+					.value();
+				if (devicesArchitectures.length > 0) {
+					let architectures = [
+						`ARCHS=${devicesArchitectures.join(" ")}`,
+						`VALID_ARCHS=${devicesArchitectures.join(" ")}`
+					];
+					if (devicesArchitectures.length > 1) {
+						architectures.push('ONLY_ACTIVE_ARCH=NO');
+					}
+					if (!buildConfig) {
+						buildConfig = {};
+					}
+					buildConfig.architectures = architectures;
+				}
+			}
+			args = args.concat((buildConfig && buildConfig.architectures) || defaultArchitectures);
+
+			args = args.concat([
+				"-sdk", "iphoneos",
+				"CONFIGURATION_BUILD_DIR=" + path.join(projectRoot, "build", "device")
+			]);
+
+			let xcodeBuildVersion = this.getXcodeVersion();
+			if (helpers.versionCompare(xcodeBuildVersion, "8.0") >= 0) {
+				this.setupAutomaticSigning(projectRoot, buildConfig).wait();
 			}
 
 			if (buildConfig && buildConfig.codeSignIdentity) {
@@ -294,51 +324,124 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 				args.push(`PROVISIONING_PROFILE=${buildConfig.mobileProvisionIdentifier}`);
 			}
 
+			if (buildConfig && buildConfig.teamIdentifier) {
+				args.push(`DEVELOPMENT_TEAM=${buildConfig.teamIdentifier}`);
+			}
+
+			// this.$logger.out("xcodebuild...");
 			this.$childProcess.spawnFromEvent("xcodebuild", args, "exit", { cwd: this.$options, stdio: 'inherit' }).wait();
+			// this.$logger.out("xcodebuild build succeded.");
 
-			if (buildForDevice) {
-				let buildOutputPath = path.join(projectRoot, "build", "device");
+			this.createIpa(projectRoot).wait();
 
-				// Produce ipa file
-				let xcrunArgs = [
-					"-sdk", "iphoneos",
-					"PackageApplication",
-					"-v", path.join(buildOutputPath, this.$projectData.projectName + ".app"),
-					"-o", path.join(buildOutputPath, this.$projectData.projectName + ".ipa")
-				];
+		}).future<void>()();
+	}
 
-				this.$childProcess.spawnFromEvent("xcrun", xcrunArgs, "exit", { cwd: this.$options, stdio: 'inherit' }).wait();
+	private setupManualSigning(projectRoot: string, buildConfig?: IiOSBuildConfig): IFuture<void> {
+		return (() => {
+			if (this.$options.provision) {
+				const pbxprojPath = path.join(projectRoot, this.$projectData.projectName + ".xcodeproj", "project.pbxproj");
+				const xcode = Xcode.open(pbxprojPath);
+				const signing = xcode.getSigning(this.$projectData.projectName);
+
+				let shouldUpdateXcode = false;
+				if (signing && signing.style === "Manual") {
+					for(let config in signing.configurations) {
+						let options = signing.configurations[config];
+						if (options.name !== this.$options.provision && options.uuid !== this.$options.provision) {
+							shouldUpdateXcode = true;
+							break;
+						}
+					}
+				} else {
+					shouldUpdateXcode = true;
+				}
+
+				if (shouldUpdateXcode) {
+					// This is slow, it read through 260 mobileprovision files on my machine and does quite some checking whether provisioning profiles and devices will match.
+					// That's why we try to avoid id by checking in the Xcode first.
+					const pickStart = Date.now();
+					const mobileprovision = this.$iOSProvisionService.pick(this.$options.provision).wait();
+					const pickEnd = Date.now();
+					this.$logger.trace("Searched and " + (mobileprovision ? "found" : "failed to find ") + " matching provisioning profile. (" + (pickEnd - pickStart) + "ms.)");
+					if (!mobileprovision) {
+						this.$errors.failWithoutHelp("Failed to find mobile provision with UUID or Name: " + this.$options.provision);
+					}
+
+					xcode.setManualSigningStyle(this.$projectData.projectName, {
+						team: mobileprovision.TeamIdentifier && mobileprovision.TeamIdentifier.length > 0 ? mobileprovision.TeamIdentifier[0] : undefined,
+						uuid: mobileprovision.UUID,
+						name: mobileprovision.Name,
+						identity: mobileprovision.Type === "Development" ? "iPhone Developer" : "iPhone Distribution"
+					});
+					xcode.save();
+
+					// this.cache(uuid);
+					this.$logger.trace("Set Manual signing style and provisioning profile.");
+				} else {
+					this.$logger.trace("The specified provisioning profile allready set in the Xcode.");
+				}
+			} else {
+				// read uuid from Xcode and cache...
 			}
 		}).future<void>()();
 	}
 
-	public buildForDeploy(platform: string, buildConfig?: IBuildConfig): IFuture<void> {
-		if (this.$options.release) {
-			return this.buildProject(this.platformData.projectRoot, buildConfig);
-		}
+	private setupAutomaticSigning(projectRoot: string, buildConfig?: IiOSBuildConfig): IFuture<void> {
+		return (() => {
+			const pbxprojPath = path.join(projectRoot, this.$projectData.projectName + ".xcodeproj", "project.pbxproj");
+			const xcode = Xcode.open(pbxprojPath);
+			const signing = xcode.getSigning(this.$projectData.projectName);
 
-		let devicesArchitectures = _(this.$devicesService.getDeviceInstances())
-			.filter(d => this.$mobileHelper.isiOSPlatform(d.deviceInfo.platform))
-			.map(d => d.deviceInfo.activeArchitecture)
-			.uniq()
-			.value();
+			if (!this.$options.provision && !(signing && signing.style === "Manual" && !this.$options.teamId)) {
+				if (buildConfig) {
+					delete buildConfig.teamIdentifier;
+				}
 
-		let architectures = [
-			`ARCHS=${devicesArchitectures.join(" ")}`,
-			`VALID_ARCHS=${devicesArchitectures.join(" ")}`
-		];
+				const teamId = this.getDevelopmentTeam();
 
-		if (devicesArchitectures.length > 1) {
-			architectures.push('ONLY_ACTIVE_ARCH=NO');
-		}
-
-		buildConfig = buildConfig || {};
-		buildConfig.architectures = architectures;
-
-		return this.buildProject(this.platformData.projectRoot, buildConfig);
+				xcode.setAutomaticSigningStyle(this.$projectData.projectName, teamId);
+				xcode.save();
+				this.$logger.trace("Set Automatic signing style and team.");
+			}
+		}).future<void>()();
 	}
 
-	public isPlatformPrepared(projectRoot: string): IFuture<boolean> {
+	private buildForSimulator(projectRoot: string, args: string[], buildConfig?: IiOSBuildConfig): IFuture<void> {
+		return (() => {
+			args = args.concat([
+				"-sdk", "iphonesimulator",
+				"ARCHS=i386 x86_64",
+				"VALID_ARCHS=i386 x86_64",
+				"ONLY_ACTIVE_ARCH=NO",
+				"CONFIGURATION_BUILD_DIR=" + path.join(projectRoot, "build", "emulator"),
+				"CODE_SIGN_IDENTITY="
+			]);
+
+			this.$childProcess.spawnFromEvent("xcodebuild", args, "exit", { cwd: this.$options, stdio: 'inherit' }).wait();
+
+		}).future<void>()();
+	}
+
+	private createIpa(projectRoot: string): IFuture<void> {
+		return (() => {
+			let buildOutputPath = path.join(projectRoot, "build", "device");
+			let xcrunArgs = [
+				"-sdk", "iphoneos",
+				"PackageApplication",
+				path.join(buildOutputPath, this.$projectData.projectName + ".app"),
+				"-o", path.join(buildOutputPath, this.$projectData.projectName + ".ipa")
+			];
+			// if (this.$logger.getLevel() !== "INFO") {
+				xcrunArgs.push("-verbose");
+			// }
+			// this.$logger.out("Packaging project...");
+			this.$childProcess.spawnFromEvent("xcrun", xcrunArgs, "exit", { cwd: this.$options, stdio: 'inherit' }).wait();
+			// this.$logger.out("Project package succeeded.");
+		}).future<void>()();
+	}
+
+	public isPlatformPrepared(projectRoot: string): boolean {
 		return this.$fs.exists(path.join(projectRoot, this.$projectData.projectName, constants.APP_FOLDER_NAME));
 	}
 
@@ -363,7 +466,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 
 			let frameworkRelativePath = '$(SRCROOT)/' + this.getLibSubpathRelativeToProjectPath(frameworkPath);
 			project.addFramework(frameworkRelativePath, frameworkAddOptions);
-			this.savePbxProj(project).wait();
+			this.savePbxProj(project);
 		}).future<void>()();
 	}
 
@@ -383,26 +486,25 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			project.addToHeaderSearchPaths({ relativePath: relativeHeaderSearchPath });
 
 			this.generateModulemap(headersSubpath, libraryName);
-			this.savePbxProj(project).wait();
+			this.savePbxProj(project);
 		}).future<void>()();
 	}
 
-	public canUpdatePlatform(installedModuleDir: string): IFuture<boolean> {
-		return (() => {
-			let currentXcodeProjectFile = this.buildPathToCurrentXcodeProjectFile();
-			let currentXcodeProjectFileContent = this.$fs.readFile(currentXcodeProjectFile).wait();
+	public canUpdatePlatform(installedModuleDir: string): boolean {
+		let currentXcodeProjectFile = this.buildPathToCurrentXcodeProjectFile();
+		let currentXcodeProjectFileContent = this.$fs.readFile(currentXcodeProjectFile);
 
-			let newXcodeProjectFile = this.buildPathToNewXcodeProjectFile(installedModuleDir);
-			this.replaceFileContent(newXcodeProjectFile).wait();
-			let newXcodeProjectFileContent = this.$fs.readFile(newXcodeProjectFile).wait();
+		let newXcodeProjectFile = this.buildPathToNewXcodeProjectFile(installedModuleDir);
+		this.replaceFileContent(newXcodeProjectFile);
+		let newXcodeProjectFileContent = this.$fs.readFile(newXcodeProjectFile);
 
-			let contentIsTheSame = currentXcodeProjectFileContent.toString() === newXcodeProjectFileContent.toString();
-			if(!contentIsTheSame) {
-				this.$logger.warn(`The content of the current project file: ${currentXcodeProjectFile} and the new project file: ${newXcodeProjectFile} is different.`);
-			}
-			return contentIsTheSame;
+		let contentIsTheSame = currentXcodeProjectFileContent.toString() === newXcodeProjectFileContent.toString();
 
-		}).future<boolean>()();
+		if (!contentIsTheSame) {
+			this.$logger.warn(`The content of the current project file: ${currentXcodeProjectFile} and the new project file: ${newXcodeProjectFile} is different.`);
+		}
+
+		return contentIsTheSame;
 	}
 
 	/**
@@ -420,7 +522,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			let platformData = this.platformData;
 			let projectPath = path.join(platformData.projectRoot, this.$projectData.projectName);
 			let projectPlist = this.getInfoPlistPath();
-			let plistContent = plist.parse(this.$fs.readText(projectPlist).wait());
+			let plistContent = plist.parse(this.$fs.readText(projectPlist));
 			let storyName = plistContent["UILaunchStoryboardName"];
 			this.$logger.trace(`Examining ${projectPlist} UILaunchStoryboardName: "${storyName}".`);
 			if (storyName !== "LaunchScreen") {
@@ -429,7 +531,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			}
 
 			let expectedStoryPath = path.join(projectPath, "Resources", "LaunchScreen.storyboard");
-			if (this.$fs.exists(expectedStoryPath).wait()) {
+			if (this.$fs.exists(expectedStoryPath)) {
 				// Found a LaunchScreen on expected path
 				this.$logger.trace("LaunchScreen.storyboard was found. Project is up to date.");
 				return;
@@ -437,7 +539,7 @@ export class IOSProjectService extends projectServiceBaseLib.PlatformProjectServ
 			this.$logger.trace("LaunchScreen file not found at: " + expectedStoryPath);
 
 			let expectedXibPath = path.join(projectPath, "en.lproj", "LaunchScreen.xib");
-			if (this.$fs.exists(expectedXibPath).wait()) {
+			if (this.$fs.exists(expectedXibPath)) {
 				this.$logger.trace("Obsolete LaunchScreen.xib was found. It'k OK, we are probably running with iOS runtime from pre v2.1.0.");
 				return;
 			}
@@ -476,8 +578,8 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
     </objects>
 </document>`;
 			try {
-				this.$fs.createDirectory(path.dirname(compatabilityXibPath)).wait();
-				this.$fs.writeFile(compatabilityXibPath, content).wait();
+				this.$fs.createDirectory(path.dirname(compatabilityXibPath));
+				this.$fs.writeFile(compatabilityXibPath, content);
 			} catch (e) {
 				this.$logger.warn("We have failed to add compatability LaunchScreen.xib due to: " + e);
 			}
@@ -486,47 +588,48 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		}
 	}
 
-	public prepareProject(): IFuture<void> {
-		return (() => {
-			let project = this.createPbxProj();
+	public prepareProject(): void {
+		if (this.$options.provision) {
+			let projectRoot = path.join(this.$projectData.platformsDir, "ios");
+			this.setupManualSigning(projectRoot).wait();
+		}
 
-			this.provideLaunchScreenIfMissing();
+		let project = this.createPbxProj();
 
-			let resources = project.pbxGroupByName("Resources");
+		this.provideLaunchScreenIfMissing();
 
-			if (resources) {
-				let references = project.pbxFileReferenceSection();
+		let resources = project.pbxGroupByName("Resources");
 
-				let xcodeProjectImages = _.map(<any[]>resources.children, resource => this.replace(references[resource.value].name));
-				this.$logger.trace("Images from Xcode project");
-				this.$logger.trace(xcodeProjectImages);
+		if (resources) {
+			let references = project.pbxFileReferenceSection();
 
-				let appResourcesImages = this.$fs.readDirectory(this.getAppResourcesDestinationDirectoryPath().wait()).wait();
-				this.$logger.trace("Current images from App_Resources");
-				this.$logger.trace(appResourcesImages);
+			let xcodeProjectImages = _.map(<any[]>resources.children, resource => this.replace(references[resource.value].name));
+			this.$logger.trace("Images from Xcode project");
+			this.$logger.trace(xcodeProjectImages);
 
-				let imagesToAdd = _.difference(appResourcesImages, xcodeProjectImages);
-				this.$logger.trace(`New images to add into xcode project: ${imagesToAdd.join(", ")}`);
-				_.each(imagesToAdd, image => project.addResourceFile(path.relative(this.platformData.projectRoot, path.join(this.getAppResourcesDestinationDirectoryPath().wait(), image))));
+			let appResourcesImages = this.$fs.readDirectory(this.getAppResourcesDestinationDirectoryPath());
+			this.$logger.trace("Current images from App_Resources");
+			this.$logger.trace(appResourcesImages);
 
-				let imagesToRemove = _.difference(xcodeProjectImages, appResourcesImages);
-				this.$logger.trace(`Images to remove from xcode project: ${imagesToRemove.join(", ")}`);
-				_.each(imagesToRemove, image => project.removeResourceFile(path.join(this.getAppResourcesDestinationDirectoryPath().wait(), image)));
+			let imagesToAdd = _.difference(appResourcesImages, xcodeProjectImages);
+			this.$logger.trace(`New images to add into xcode project: ${imagesToAdd.join(", ")}`);
+			_.each(imagesToAdd, image => project.addResourceFile(path.relative(this.platformData.projectRoot, path.join(this.getAppResourcesDestinationDirectoryPath(), image))));
 
-				this.savePbxProj(project).wait();
-			}
-		}).future<void>()();
+			let imagesToRemove = _.difference(xcodeProjectImages, appResourcesImages);
+			this.$logger.trace(`Images to remove from xcode project: ${imagesToRemove.join(", ")}`);
+			_.each(imagesToRemove, image => project.removeResourceFile(path.join(this.getAppResourcesDestinationDirectoryPath(), image)));
+
+			this.savePbxProj(project);
+		}
 	}
 
-	public prepareAppResources(appResourcesDirectoryPath: string): IFuture<void> {
-		return (() => {
-			let platformFolder = path.join(appResourcesDirectoryPath, this.platformData.normalizedPlatformName);
-			let filterFile = (filename: string) => this.$fs.deleteFile(path.join(platformFolder, filename)).wait();
+	public prepareAppResources(appResourcesDirectoryPath: string): void {
+		let platformFolder = path.join(appResourcesDirectoryPath, this.platformData.normalizedPlatformName);
+		let filterFile = (filename: string) => this.$fs.deleteFile(path.join(platformFolder, filename));
 
-			filterFile(this.platformData.configurationFileName);
+		filterFile(this.platformData.configurationFileName);
 
-			this.$fs.deleteDirectory(this.getAppResourcesDestinationDirectoryPath().wait()).wait();
-		}).future<void>()();
+		this.$fs.deleteDirectory(this.getAppResourcesDestinationDirectoryPath());
 	}
 
 	public processConfigurationFilesFromAppResources(): IFuture<void> {
@@ -536,7 +639,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 			_(this.getAllInstalledPlugins().wait())
 				.map(pluginData => this.$pluginVariablesService.interpolatePluginVariables(pluginData, this.platformData.configurationFilePath).wait())
 				.value();
-			this.$pluginVariablesService.interpolateAppIdentifier(this.platformData.configurationFilePath).wait();
+			this.$pluginVariablesService.interpolateAppIdentifier(this.platformData.configurationFilePath);
 		}).future<void>()();
 	}
 
@@ -550,7 +653,12 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 				this.platformData.configurationFileName
 			);
 	}
-	public ensureConfigurationFileInAppResources(): IFuture<void> {
+
+	public ensureConfigurationFileInAppResources(): void {
+		return null;
+	}
+
+	public cleanProject(projectRoot: string, options: string[]): IFuture<void> {
 		return Future.fromResult();
 	}
 
@@ -558,16 +666,16 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		return (() => {
 			let projectDir = this.$projectData.projectDir;
 			let infoPlistPath = this.$options.baseConfig || path.join(projectDir, constants.APP_FOLDER_NAME, constants.APP_RESOURCES_FOLDER_NAME, this.platformData.normalizedPlatformName, this.platformData.configurationFileName);
-			this.ensureConfigurationFileInAppResources().wait();
+			this.ensureConfigurationFileInAppResources();
 
-			if (!this.$fs.exists(infoPlistPath).wait()) {
+			if (!this.$fs.exists(infoPlistPath)) {
 				this.$logger.trace("Info.plist: No app/App_Resources/iOS/Info.plist found, falling back to pre-1.6.0 Info.plist behavior.");
 				return;
 			}
 
 			let session = new PlistSession({ log: (txt: string) => this.$logger.trace("Info.plist: " + txt) });
 			let makePatch = (plistPath: string) => {
-				if (!this.$fs.exists(plistPath).wait()) {
+				if (!this.$fs.exists(plistPath)) {
 					this.$logger.trace("No plist found at: " + plistPath);
 					return;
 				}
@@ -575,7 +683,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 				this.$logger.trace("Schedule merge plist at: " + plistPath);
 				session.patch({
 					name: path.relative(projectDir, plistPath),
-					read: () => this.$fs.readFile(plistPath).wait().toString()
+					read: () => this.$fs.readText(plistPath)
 				});
 			};
 
@@ -605,7 +713,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 			let plistContent = session.build();
 
 			this.$logger.trace("Info.plist: Write to: " + this.platformData.configurationFilePath);
-			this.$fs.writeFile(this.platformData.configurationFilePath, plistContent).wait();
+			this.$fs.writeFile(this.platformData.configurationFilePath, plistContent);
 
 		}).future<void>()();
 	}
@@ -654,7 +762,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		return project;
 	}
 
-	private savePbxProj(project: any): IFuture<void> {
+	private savePbxProj(project: any): void {
 		return this.$fs.writeFile(this.pbxProjPath, project.writeSync());
 	}
 
@@ -664,36 +772,34 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 
 			this.prepareFrameworks(pluginPlatformsFolderPath, pluginData).wait();
 			this.prepareStaticLibs(pluginPlatformsFolderPath, pluginData).wait();
-			this.prepareCocoapods(pluginPlatformsFolderPath).wait();
+			this.prepareCocoapods(pluginPlatformsFolderPath);
 		}).future<void>()();
 	}
 
-	public removePluginNativeCode(pluginData: IPluginData): IFuture<void> {
-		return (() => {
-			let pluginPlatformsFolderPath = pluginData.pluginPlatformsFolderPath(IOSProjectService.IOS_PLATFORM_NAME);
+	public removePluginNativeCode(pluginData: IPluginData): void {
+		let pluginPlatformsFolderPath = pluginData.pluginPlatformsFolderPath(IOSProjectService.IOS_PLATFORM_NAME);
 
-			this.removeFrameworks(pluginPlatformsFolderPath, pluginData).wait();
-			this.removeStaticLibs(pluginPlatformsFolderPath, pluginData).wait();
-			this.removeCocoapods(pluginPlatformsFolderPath).wait();
-		}).future<void>()();
+		this.removeFrameworks(pluginPlatformsFolderPath, pluginData);
+		this.removeStaticLibs(pluginPlatformsFolderPath, pluginData);
+		this.removeCocoapods(pluginPlatformsFolderPath);
 	}
 
 	public afterPrepareAllPlugins(): IFuture<void> {
 		return (() => {
-			if (this.$fs.exists(this.projectPodFilePath).wait()) {
-				let projectPodfileContent = this.$fs.readText(this.projectPodFilePath).wait();
+			if (this.$fs.exists(this.projectPodFilePath)) {
+				let projectPodfileContent = this.$fs.readText(this.projectPodFilePath);
 				this.$logger.trace("Project Podfile content");
 				this.$logger.trace(projectPodfileContent);
 
 				let firstPostInstallIndex = projectPodfileContent.indexOf(IOSProjectService.PODFILE_POST_INSTALL_SECTION_NAME);
 				if (firstPostInstallIndex !== -1 && firstPostInstallIndex !== projectPodfileContent.lastIndexOf(IOSProjectService.PODFILE_POST_INSTALL_SECTION_NAME)) {
-					this.$cocoapodsService.mergePodfileHookContent(IOSProjectService.PODFILE_POST_INSTALL_SECTION_NAME, this.projectPodFilePath).wait();
+					this.$cocoapodsService.mergePodfileHookContent(IOSProjectService.PODFILE_POST_INSTALL_SECTION_NAME, this.projectPodFilePath);
 				}
 
 				let xcuserDataPath = path.join(this.xcodeprojPath, "xcuserdata");
 				let sharedDataPath = path.join(this.xcodeprojPath, "xcshareddata");
 
-				if (!this.$fs.exists(xcuserDataPath).wait() && !this.$fs.exists(sharedDataPath).wait()) {
+				if (!this.$fs.exists(xcuserDataPath) && !this.$fs.exists(sharedDataPath)) {
 					this.$logger.info("Creating project scheme...");
 
 					this.checkIfXcodeprojIsRequired().wait();
@@ -711,7 +817,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		return Future.fromResult();
 	}
 
-	private getAllLibsForPluginWithFileExtension(pluginData: IPluginData, fileExtension: string): IFuture<string[]> {
+	private getAllLibsForPluginWithFileExtension(pluginData: IPluginData, fileExtension: string): string[] {
 		let filterCallback = (fileName: string, pluginPlatformsFolderPath: string) => path.extname(fileName) === fileExtension;
 		return this.getAllNativeLibrariesForPlugin(pluginData, IOSProjectService.IOS_PLATFORM_NAME, filterCallback);
 	};
@@ -727,7 +833,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 	private validateFramework(libraryPath: string): IFuture<void> {
 		return (() => {
 			let infoPlistPath = path.join(libraryPath, "Info.plist");
-			if (!this.$fs.exists(infoPlistPath).wait()) {
+			if (!this.$fs.exists(infoPlistPath)) {
 				this.$errors.failWithoutHelp("The bundle at %s does not contain an Info.plist file.", libraryPath);
 			}
 
@@ -755,21 +861,17 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		}).future<void>()();
 	}
 
-	private replaceFileContent(file: string): IFuture<void> {
-		return (() => {
-			let fileContent = this.$fs.readText(file).wait();
-			let replacedContent = helpers.stringReplaceAll(fileContent, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER, this.$projectData.projectName);
-			this.$fs.writeFile(file, replacedContent).wait();
-		}).future<void>()();
+	private replaceFileContent(file: string): void {
+		let fileContent = this.$fs.readText(file);
+		let replacedContent = helpers.stringReplaceAll(fileContent, IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER, this.$projectData.projectName);
+		this.$fs.writeFile(file, replacedContent);
 	}
 
-	private replaceFileName(fileNamePart: string, fileRootLocation: string): IFuture<void> {
-		return (() => {
-			let oldFileName = IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + fileNamePart;
-			let newFileName = this.$projectData.projectName + fileNamePart;
+	private replaceFileName(fileNamePart: string, fileRootLocation: string): void {
+		let oldFileName = IOSProjectService.IOS_PROJECT_NAME_PLACEHOLDER + fileNamePart;
+		let newFileName = this.$projectData.projectName + fileNamePart;
 
-			this.$fs.rename(path.join(fileRootLocation, oldFileName), path.join(fileRootLocation, newFileName)).wait();
-		}).future<void>()();
+		this.$fs.rename(path.join(fileRootLocation, oldFileName), path.join(fileRootLocation, newFileName));
 	}
 
 	private executePodInstall(): IFuture<any> {
@@ -813,96 +915,88 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 
 	private prepareFrameworks(pluginPlatformsFolderPath: string, pluginData: IPluginData): IFuture<void> {
 		return (() => {
-			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".framework").wait(), fileName => this.addFramework(path.join(pluginPlatformsFolderPath, fileName)).wait());
+			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".framework"), fileName => this.addFramework(path.join(pluginPlatformsFolderPath, fileName)).wait());
 		}).future<void>()();
 	}
 
 	private prepareStaticLibs(pluginPlatformsFolderPath: string, pluginData: IPluginData): IFuture<void> {
 		return (() => {
-			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".a").wait(), fileName => this.addStaticLibrary(path.join(pluginPlatformsFolderPath, fileName)).wait());
+			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".a"), fileName => this.addStaticLibrary(path.join(pluginPlatformsFolderPath, fileName)).wait());
 		}).future<void>()();
 	}
 
-	private prepareCocoapods(pluginPlatformsFolderPath: string, opts?: any): IFuture<void> {
-		return (() => {
-			let pluginPodFilePath = path.join(pluginPlatformsFolderPath, "Podfile");
-			if (this.$fs.exists(pluginPodFilePath).wait()) {
-				let pluginPodFileContent = this.$fs.readText(pluginPodFilePath).wait(),
-					pluginPodFilePreparedContent = this.buildPodfileContent(pluginPodFilePath, pluginPodFileContent),
-					projectPodFileContent = this.$fs.exists(this.projectPodFilePath).wait() ? this.$fs.readText(this.projectPodFilePath).wait() : "";
+	private prepareCocoapods(pluginPlatformsFolderPath: string, opts?: any): void {
+		let pluginPodFilePath = path.join(pluginPlatformsFolderPath, "Podfile");
+		if (this.$fs.exists(pluginPodFilePath)) {
+			let pluginPodFileContent = this.$fs.readText(pluginPodFilePath),
+				pluginPodFilePreparedContent = this.buildPodfileContent(pluginPodFilePath, pluginPodFileContent),
+				projectPodFileContent = this.$fs.exists(this.projectPodFilePath) ? this.$fs.readText(this.projectPodFilePath) : "";
 
-				if (!~projectPodFileContent.indexOf(pluginPodFilePreparedContent)) {
-					let podFileHeader = this.$cocoapodsService.getPodfileHeader(this.$projectData.projectName),
-						podFileFooter = this.$cocoapodsService.getPodfileFooter();
+			if (!~projectPodFileContent.indexOf(pluginPodFilePreparedContent)) {
+				let podFileHeader = this.$cocoapodsService.getPodfileHeader(this.$projectData.projectName),
+					podFileFooter = this.$cocoapodsService.getPodfileFooter();
 
-					if (_.startsWith(projectPodFileContent, podFileHeader)) {
-						projectPodFileContent = projectPodFileContent.substr(podFileHeader.length);
-					}
-
-					if (_.endsWith(projectPodFileContent, podFileFooter)) {
-						projectPodFileContent = projectPodFileContent.substr(0, projectPodFileContent.length - podFileFooter.length);
-					}
-
-					let contentToWrite = `${podFileHeader}${projectPodFileContent}${pluginPodFilePreparedContent}${podFileFooter}`;
-					this.$fs.writeFile(this.projectPodFilePath, contentToWrite).wait();
-
-					let project = this.createPbxProj();
-					this.savePbxProj(project).wait();
-				}
-			}
-
-			if (opts && opts.executePodInstall && this.$fs.exists(pluginPodFilePath).wait()) {
-				this.executePodInstall().wait();
-			}
-		}).future<void>()();
-	}
-
-	private removeFrameworks(pluginPlatformsFolderPath: string, pluginData: IPluginData): IFuture<void> {
-		return (() => {
-			let project = this.createPbxProj();
-			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".framework").wait(), fileName => {
-				let relativeFrameworkPath = this.getLibSubpathRelativeToProjectPath(fileName);
-				project.removeFramework(relativeFrameworkPath, { customFramework: true, embed: true });
-			});
-
-			this.savePbxProj(project).wait();
-		}).future<void>()();
-	}
-
-	private removeStaticLibs(pluginPlatformsFolderPath: string, pluginData: IPluginData): IFuture<void> {
-		return (() => {
-			let project = this.createPbxProj();
-
-			_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".a").wait(), fileName => {
-				let staticLibPath = path.join(pluginPlatformsFolderPath, fileName);
-				let relativeStaticLibPath = this.getLibSubpathRelativeToProjectPath(path.basename(staticLibPath));
-				project.removeFramework(relativeStaticLibPath);
-
-				let headersSubpath = path.join("include", path.basename(staticLibPath, ".a"));
-				let relativeHeaderSearchPath = path.join(this.getLibSubpathRelativeToProjectPath(headersSubpath));
-				project.removeFromHeaderSearchPaths({ relativePath: relativeHeaderSearchPath });
-			});
-
-			this.savePbxProj(project).wait();
-		}).future<void>()();
-	}
-
-	private removeCocoapods(pluginPlatformsFolderPath: string): IFuture<void> {
-		return (() => {
-			let pluginPodFilePath = path.join(pluginPlatformsFolderPath, "Podfile");
-			if (this.$fs.exists(pluginPodFilePath).wait() && this.$fs.exists(this.projectPodFilePath).wait()) {
-				let pluginPodFileContent = this.$fs.readText(pluginPodFilePath).wait();
-				let projectPodFileContent = this.$fs.readText(this.projectPodFilePath).wait();
-				let contentToRemove = this.buildPodfileContent(pluginPodFilePath, pluginPodFileContent);
-				projectPodFileContent = helpers.stringReplaceAll(projectPodFileContent, contentToRemove, "");
-				if (projectPodFileContent.trim() === `use_frameworks!${os.EOL}${os.EOL}target "${this.$projectData.projectName}" do${os.EOL}${os.EOL}end`) {
-					this.$fs.deleteFile(this.projectPodFilePath).wait();
-				} else {
-					this.$fs.writeFile(this.projectPodFilePath, projectPodFileContent).wait();
+				if (_.startsWith(projectPodFileContent, podFileHeader)) {
+					projectPodFileContent = projectPodFileContent.substr(podFileHeader.length);
 				}
 
+				if (_.endsWith(projectPodFileContent, podFileFooter)) {
+					projectPodFileContent = projectPodFileContent.substr(0, projectPodFileContent.length - podFileFooter.length);
+				}
+
+				let contentToWrite = `${podFileHeader}${projectPodFileContent}${pluginPodFilePreparedContent}${podFileFooter}`;
+				this.$fs.writeFile(this.projectPodFilePath, contentToWrite);
+
+				let project = this.createPbxProj();
+				this.savePbxProj(project);
 			}
-		}).future<void>()();
+		}
+
+		if (opts && opts.executePodInstall && this.$fs.exists(pluginPodFilePath)) {
+			this.executePodInstall().wait();
+		}
+	}
+
+	private removeFrameworks(pluginPlatformsFolderPath: string, pluginData: IPluginData): void {
+		let project = this.createPbxProj();
+		_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".framework"), fileName => {
+			let relativeFrameworkPath = this.getLibSubpathRelativeToProjectPath(fileName);
+			project.removeFramework(relativeFrameworkPath, { customFramework: true, embed: true });
+		});
+
+		this.savePbxProj(project);
+	}
+
+	private removeStaticLibs(pluginPlatformsFolderPath: string, pluginData: IPluginData): void {
+		let project = this.createPbxProj();
+
+		_.each(this.getAllLibsForPluginWithFileExtension(pluginData, ".a"), fileName => {
+			let staticLibPath = path.join(pluginPlatformsFolderPath, fileName);
+			let relativeStaticLibPath = this.getLibSubpathRelativeToProjectPath(path.basename(staticLibPath));
+			project.removeFramework(relativeStaticLibPath);
+
+			let headersSubpath = path.join("include", path.basename(staticLibPath, ".a"));
+			let relativeHeaderSearchPath = path.join(this.getLibSubpathRelativeToProjectPath(headersSubpath));
+			project.removeFromHeaderSearchPaths({ relativePath: relativeHeaderSearchPath });
+		});
+
+		this.savePbxProj(project);
+	}
+
+	private removeCocoapods(pluginPlatformsFolderPath: string): void {
+		let pluginPodFilePath = path.join(pluginPlatformsFolderPath, "Podfile");
+
+		if (this.$fs.exists(pluginPodFilePath) && this.$fs.exists(this.projectPodFilePath)) {
+			let pluginPodFileContent = this.$fs.readText(pluginPodFilePath);
+			let projectPodFileContent = this.$fs.readText(this.projectPodFilePath);
+			let contentToRemove = this.buildPodfileContent(pluginPodFilePath, pluginPodFileContent);
+			projectPodFileContent = helpers.stringReplaceAll(projectPodFileContent, contentToRemove, "");
+			if (projectPodFileContent.trim() === `use_frameworks!${os.EOL}${os.EOL}target "${this.$projectData.projectName}" do${os.EOL}${os.EOL}end`) {
+				this.$fs.deleteFile(this.projectPodFilePath);
+			} else {
+				this.$fs.writeFile(this.projectPodFilePath, projectPodFileContent);
+			}
+		}
 	}
 
 	private buildPodfileContent(pluginPodFilePath: string, pluginPodFileContent: string): string {
@@ -910,25 +1004,25 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 	}
 
 	private generateModulemap(headersFolderPath: string, libraryName: string): void {
-		let headersFilter = (fileName: string, containingFolderPath: string) => (path.extname(fileName) === ".h" && this.$fs.getFsStats(path.join(containingFolderPath, fileName)).wait().isFile());
-		let headersFolderContents = this.$fs.readDirectory(headersFolderPath).wait();
+		let headersFilter = (fileName: string, containingFolderPath: string) => (path.extname(fileName) === ".h" && this.$fs.getFsStats(path.join(containingFolderPath, fileName)).isFile());
+		let headersFolderContents = this.$fs.readDirectory(headersFolderPath);
 		let headers = _(headersFolderContents).filter(item => headersFilter(item, headersFolderPath)).value();
 
 		if (!headers.length) {
-			this.$fs.deleteFile(path.join(headersFolderPath, "module.modulemap")).wait();
+			this.$fs.deleteFile(path.join(headersFolderPath, "module.modulemap"));
 			return;
 		}
 
 		headers = _.map(headers, value => `header "${value}"`);
 
 		let modulemap = `module ${libraryName} { explicit module ${libraryName} { ${headers.join(" ")} } }`;
-		this.$fs.writeFile(path.join(headersFolderPath, "module.modulemap"), modulemap).wait();
+		this.$fs.writeFile(path.join(headersFolderPath, "module.modulemap"), modulemap);
 	}
 
 	private mergeXcconfigFiles(pluginFile: string, projectFile: string): IFuture<void> {
 		return (() => {
-			if (!this.$fs.exists(projectFile).wait()) {
-				this.$fs.writeFile(projectFile, "").wait();
+			if (!this.$fs.exists(projectFile)) {
+				this.$fs.writeFile(projectFile, "");
 			}
 
 			this.checkIfXcodeprojIsRequired().wait();
@@ -941,25 +1035,25 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 
 	private mergeProjectXcconfigFiles(): IFuture<void> {
 		return (() => {
-			this.$fs.deleteFile(this.$options.release ? this.pluginsReleaseXcconfigFilePath : this.pluginsDebugXcconfigFilePath).wait();
+			this.$fs.deleteFile(this.$options.release ? this.pluginsReleaseXcconfigFilePath : this.pluginsDebugXcconfigFilePath);
 
 			let allPlugins: IPluginData[] = (<IPluginsService>this.$injector.resolve("pluginsService")).getAllInstalledPlugins().wait();
 			for (let plugin of allPlugins) {
 				let pluginPlatformsFolderPath = plugin.pluginPlatformsFolderPath(IOSProjectService.IOS_PLATFORM_NAME);
 				let pluginXcconfigFilePath = path.join(pluginPlatformsFolderPath, "build.xcconfig");
-				if (this.$fs.exists(pluginXcconfigFilePath).wait()) {
+				if (this.$fs.exists(pluginXcconfigFilePath)) {
 					this.mergeXcconfigFiles(pluginXcconfigFilePath, this.$options.release ? this.pluginsReleaseXcconfigFilePath : this.pluginsDebugXcconfigFilePath).wait();
 				}
 			}
 
 			let appResourcesXcconfigPath = path.join(this.$projectData.projectDir, constants.APP_FOLDER_NAME, constants.APP_RESOURCES_FOLDER_NAME, this.platformData.normalizedPlatformName, "build.xcconfig");
-			if (this.$fs.exists(appResourcesXcconfigPath).wait()) {
+			if (this.$fs.exists(appResourcesXcconfigPath)) {
 				this.mergeXcconfigFiles(appResourcesXcconfigPath, this.$options.release ? this.pluginsReleaseXcconfigFilePath : this.pluginsDebugXcconfigFilePath).wait();
 			}
 
 			let podFilesRootDirName = path.join("Pods", "Target Support Files", `Pods-${this.$projectData.projectName}`);
 			let podFolder = path.join(this.platformData.projectRoot, podFilesRootDirName);
-			if (this.$fs.exists(podFolder).wait()) {
+			if (this.$fs.exists(podFolder)) {
 				if (this.$options.release) {
 					this.mergeXcconfigFiles(path.join(this.platformData.projectRoot, podFilesRootDirName, `Pods-${this.$projectData.projectName}.release.xcconfig`), this.pluginsReleaseXcconfigFilePath).wait();
 				} else {
@@ -999,22 +1093,22 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		return xcodeBuildVersion;
 	}
 
-	private getDevelopmentTeams(): Array<{ id: string, name: string }> {
+	private getDevelopmentTeams(): Array<{id:string, name: string}> {
 		let dir = path.join(process.env.HOME, "Library/MobileDevice/Provisioning Profiles/");
-		let files = this.$fs.readDirectory(dir).wait();
+		let files = this.$fs.readDirectory(dir);
 		let teamIds: any = {};
 		for (let file of files) {
 			let filePath = path.join(dir, file);
-			let data = this.$fs.readText(filePath, "utf8").wait();
+			let data = this.$fs.readText(filePath, "utf8");
 			let teamId = this.getProvisioningProfileValue("TeamIdentifier", data);
 			let teamName = this.getProvisioningProfileValue("TeamName", data);
 			if (teamId) {
 				teamIds[teamId] = teamName;
 			}
 		}
-		let teamIdsArray = new Array<{ id: string, name: string }>();
+		let teamIdsArray = new Array<{id:string, name: string}>();
 		for (let teamId in teamIds) {
-			teamIdsArray.push({ id: teamId, name: teamIds[teamId] });
+			teamIdsArray.push({ id:teamId, name:teamIds[teamId] });
 		}
 		return teamIdsArray;
 	}
@@ -1023,7 +1117,7 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 		let findStr = "<key>" + name + "</key>";
 		let index = text.indexOf(findStr);
 		if (index > 0) {
-			index = text.indexOf("<string>", index + findStr.length);
+			index = text.indexOf("<string>", index+findStr.length);
 			if (index > 0) {
 				index += "<string>".length;
 				let endIndex = text.indexOf("</string>", index);
@@ -1036,14 +1130,14 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 
 	private readTeamId(): string {
 		let xcconfigFile = path.join(this.$projectData.appResourcesDirectoryPath, this.platformData.normalizedPlatformName, "build.xcconfig");
-		if (this.$fs.exists(xcconfigFile).wait()) {
-			let text = this.$fs.readText(xcconfigFile).wait();
+		if (this.$fs.exists(xcconfigFile)) {
+			let text = this.$fs.readText(xcconfigFile);
 			let teamId: string;
 			text.split(/\r?\n/).forEach((line) => {
 				line = line.replace(/\/(\/)[^\n]*$/, "");
 				if (line.indexOf("DEVELOPMENT_TEAM") >= 0) {
 					teamId = line.split("=")[1].trim();
-					if (teamId[teamId.length - 1] === ';') {
+					if (teamId[teamId.length-1] === ';') {
 						teamId = teamId.slice(0, -1);
 					}
 				}
@@ -1053,8 +1147,8 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 			}
 		}
 		let fileName = path.join(this.platformData.projectRoot, "teamid");
-		if (this.$fs.exists(fileName).wait()) {
-			return this.$fs.readText(fileName).wait();
+		if (this.$fs.exists(fileName)) {
+			return this.$fs.readText(fileName);
 		}
 		return null;
 	}
@@ -1086,11 +1180,11 @@ We will now place an empty obsolete compatability white screen LauncScreen.xib f
 					"Yes, persist the team id in platforms folder.",
 					"No, don't persist this setting."
 				];
-				let choicePersist = this.$prompter.promptForChoice("Do you want to make teamId: " + teamId + " a persistent choice for your app?", choicesPersist).wait();
+				let choicePersist = this.$prompter.promptForChoice("Do you want to make teamId: "+teamId+" a persistent choice for your app?", choicesPersist).wait();
 				switch (choicesPersist.indexOf(choicePersist)) {
 					case 0:
 						let xcconfigFile = path.join(this.$projectData.appResourcesDirectoryPath, this.platformData.normalizedPlatformName, "build.xcconfig");
-						this.$fs.appendFile(xcconfigFile, "\nDEVELOPMENT_TEAM = " + teamId + "\n").wait();
+						this.$fs.appendFile(xcconfigFile, "\nDEVELOPMENT_TEAM = " + teamId + "\n");
 						break;
 					case 1:
 						this.$fs.writeFile(path.join(this.platformData.projectRoot, "teamid"), teamId);
