@@ -3,11 +3,11 @@ import * as helpers from "../../common/helpers";
 import * as path from "path";
 import * as semver from "semver";
 import * as fiberBootstrap from "../../common/fiber-bootstrap";
+import { NodeModulesDependenciesBuilder } from "../../tools/node-modules/node-modules-dependencies-builder";
 
-let gaze = require("gaze");
+let choki = require("chokidar");
 
 class LiveSyncService implements ILiveSyncService {
-	public forceExecuteFullSync = false;
 	private _isInitialized = false;
 
 	constructor(private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
@@ -30,7 +30,7 @@ class LiveSyncService implements ILiveSyncService {
 	private ensureAndroidFrameworkVersion(platformData: IPlatformData): IFuture<void> { // TODO: this can be moved inside command or canExecute function
 		return (() => {
 			this.$projectDataService.initialize(this.$projectData.projectDir);
-			let frameworkVersion = this.$projectDataService.getValue(platformData.frameworkPackageName).wait().version;
+			let frameworkVersion = this.$projectDataService.getValue(platformData.frameworkPackageName).version;
 
 			if (platformData.normalizedPlatformName.toLowerCase() === this.$devicePlatformsConstants.Android.toLowerCase()) {
 				if (semver.lt(frameworkVersion, "1.2.1")) {
@@ -49,8 +49,11 @@ class LiveSyncService implements ILiveSyncService {
 		return this._isInitialized;
 	}
 
-	public liveSync(platform: string, applicationReloadAction?: (deviceAppData: Mobile.IDeviceAppData, localToDevicePaths: Mobile.ILocalToDevicePathData[]) => IFuture<void>): IFuture<void> {
+	public liveSync(platform: string, applicationReloadAction?: (deviceAppData: Mobile.IDeviceAppData) => IFuture<void>): IFuture<void> {
 		return (() => {
+			if (this.$options.justlaunch) {
+				this.$options.watch = false;
+			}
 			let liveSyncData: ILiveSyncData[] = [];
 			if (platform) {
 				this.$devicesService.initialize({ platform: platform, deviceId: this.$options.device  }).wait();
@@ -62,7 +65,7 @@ class LiveSyncService implements ILiveSyncService {
 			} else {
 				this.$devicesService.initialize({ skipInferPlatform: true }).wait();
 				this.$devicesService.stopDeviceDetectionInterval().wait();
-				for(let installedPlatform of this.$platformService.getInstalledPlatforms().wait()) {
+				for(let installedPlatform of this.$platformService.getInstalledPlatforms()) {
 					if (this.$devicesService.getDevicesForPlatform(installedPlatform).length === 0) {
 						this.$devicesService.startEmulator(installedPlatform).wait();
 					}
@@ -80,7 +83,6 @@ class LiveSyncService implements ILiveSyncService {
 
 	private prepareLiveSyncData(platform: string): ILiveSyncData {
 		platform = platform || this.$devicesService.platform;
-		this.$platformService.preparePlatform(platform.toLowerCase()).wait();
 		let platformData = this.$platformsData.getPlatformData(platform.toLowerCase());
 		if (this.$mobileHelper.isAndroidPlatform(platform)) {
 			this.ensureAndroidFrameworkVersion(platformData).wait();
@@ -89,31 +91,28 @@ class LiveSyncService implements ILiveSyncService {
 			platform: platform,
 			appIdentifier: this.$projectData.projectId,
 			projectFilesPath: path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME),
-			syncWorkingDirectory: path.join(this.$projectData.projectDir, constants.APP_FOLDER_NAME),
-			excludedProjectDirsAndFiles: this.$options.release ? constants.LIVESYNC_EXCLUDED_FILE_PATTERNS : [],
-			forceExecuteFullSync: this.forceExecuteFullSync
+			syncWorkingDirectory: this.$projectData.projectDir,
+			excludedProjectDirsAndFiles: this.$options.release ? constants.LIVESYNC_EXCLUDED_FILE_PATTERNS : []
 		};
 
 		return liveSyncData;
 	}
 
-	private resolvePlatformLiveSyncBaseService(platform: string, liveSyncData: ILiveSyncData): IPlatformLiveSyncService {
-		return this.$injector.resolve(this.$liveSyncProvider.platformSpecificLiveSyncServices[platform.toLowerCase()], { _liveSyncData: liveSyncData });
-	}
-
 	@helpers.hook('livesync')
 	private liveSyncCore(liveSyncData: ILiveSyncData[], applicationReloadAction: (deviceAppData: Mobile.IDeviceAppData, localToDevicePaths: Mobile.ILocalToDevicePathData[]) => IFuture<void>): IFuture<void> {
 		return (() => {
+			this.$platformService.trackProjectType().wait();
+
 			let watchForChangeActions: ((event: string, filePath: string, dispatcher: IFutureDispatcher) => void)[] = [];
 			_.each(liveSyncData, (dataItem) => {
-				let service = this.resolvePlatformLiveSyncBaseService(dataItem.platform, dataItem);
+				let service: IPlatformLiveSyncService = this.$injector.resolve("platformLiveSyncService", { _liveSyncData: dataItem });
 				watchForChangeActions.push((event: string, filePath: string, dispatcher: IFutureDispatcher) => {
 					service.partialSync(event, filePath, dispatcher, applicationReloadAction);
 				});
 				service.fullSync(applicationReloadAction).wait();
 			});
 
-			if(this.$options.watch) {
+			if(this.$options.watch && !this.$options.justlaunch) {
 				this.$hooksService.executeBeforeHooks('watch').wait();
 				this.partialSync(liveSyncData[0].syncWorkingDirectory, watchForChangeActions);
 			}
@@ -122,25 +121,39 @@ class LiveSyncService implements ILiveSyncService {
 
 	private partialSync(syncWorkingDirectory: string, onChangedActions: ((event: string, filePath: string, dispatcher: IFutureDispatcher) => void )[]): void {
 		let that = this;
+		let dependenciesBuilder = this.$injector.resolve(NodeModulesDependenciesBuilder, {});
+		let productionDependencies = dependenciesBuilder.getProductionDependencies(this.$projectData.projectDir);
+		let pattern = ["app"];
 
-		let gazeWatcher = gaze("**/*", { cwd: syncWorkingDirectory }, function (err: any, watcher: any) {
-			this.on('all', (event: string, filePath: string) => {
-				fiberBootstrap.run(() => {
-					that.$dispatcher.dispatch(() => (() => {
-						try {
-							for (let i = 0; i < onChangedActions.length; i++) {
-								onChangedActions[i](event, filePath, that.$dispatcher);
-							}
-						} catch (err) {
-							that.$logger.info(`Unable to sync file ${filePath}. Error is:${err.message}`.red.bold);
-							that.$logger.info("Try saving it again or restart the livesync operation.");
+		if(this.$options.syncAllFiles) {
+			pattern.push("package.json");
+
+			// watch only production node_module/packages same one prepare uses
+			for(let index in productionDependencies) {
+				pattern.push("node_modules/" + productionDependencies[index].name);
+			}
+		}
+
+		let watcher = choki.watch(pattern, { ignoreInitial: true, cwd: syncWorkingDirectory, ignored: '**/*.DS_Store' }).on("all", (event: string, filePath: string) => {
+			fiberBootstrap.run(() => {
+				that.$dispatcher.dispatch(() => (() => {
+					try {
+						that.$logger.trace(`Event '${event}' triggered for path: '${filePath}'`);
+						filePath = path.join(syncWorkingDirectory, filePath);
+						for (let i = 0; i < onChangedActions.length; i++) {
+							onChangedActions[i](event, filePath, that.$dispatcher);
 						}
-					}).future<void>()());
-				});
+					} catch (err) {
+						that.$logger.info(`Unable to sync file ${filePath}. Error is:${err.message}`.red.bold);
+						that.$logger.info("Try saving it again or restart the livesync operation.");
+					}
+				}).future<void>()());
 			});
 		});
 
-		this.$processService.attachToProcessExitSignals(this, () => gazeWatcher.close());
+		this.$processService.attachToProcessExitSignals(this, () => {
+			watcher.close(pattern);
+		});
 		this.$dispatcher.run();
 	}
 }
