@@ -20,61 +20,50 @@ export class ProjectDataService implements IProjectDataService {
 		}
 	}
 
-	public getValue(propertyName: string): IFuture<any> {
-		return (() => {
-			this.loadProjectFile().wait();
-			return this.projectData ? this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][propertyName] : null;
-		}).future<string>()();
+	public getValue(propertyName: string): any {
+		this.loadProjectFile();
+		return this.projectData ? this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][propertyName] : null;
 	}
 
-	public setValue(key: string, value: any): IFuture<void> {
-		return (() => {
-			this.loadProjectFile().wait();
-			if (!this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE]) {
-				this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE] = Object.create(null);
-			}
-			this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][key] = value;
-			this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent).wait();
-		}).future<void>()();
+	public setValue(key: string, value: any): void {
+		this.loadProjectFile();
+		if (!this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE]) {
+			this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE] = Object.create(null);
+		}
+		this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][key] = value;
+		this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent);
 	}
 
-	public removeProperty(propertyName: string): IFuture<void> {
-		return (() => {
-			this.loadProjectFile().wait();
-			delete this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][propertyName];
-			this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent).wait();
-		}).future<void>()();
+	public removeProperty(propertyName: string): void {
+		this.loadProjectFile();
+		delete this.projectData[this.$staticConfig.CLIENT_NAME_KEY_IN_PROJECT_FILE][propertyName];
+		this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent);
 	}
 
-	public removeDependency(dependencyName: string): IFuture<void> {
-		return (() => {
-			this.loadProjectFile().wait();
-			delete this.projectData[ProjectDataService.DEPENDENCIES_KEY_NAME][dependencyName];
-			this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent).wait();
-		}).future<void>()();
+	public removeDependency(dependencyName: string): void {
+		this.loadProjectFile();
+		delete this.projectData[ProjectDataService.DEPENDENCIES_KEY_NAME][dependencyName];
+		this.$fs.writeJson(this.projectFilePath, this.projectData, this.projectFileIndent);
 	}
 
-	private loadProjectFile(): IFuture<void> {
-		return (() => {
-			assert.ok(this.projectFilePath, "Initialize method of projectDataService is not called.");
+	private loadProjectFile(): void {
+		assert.ok(this.projectFilePath, "Initialize method of projectDataService is not called.");
 
-			if (!this.$fs.exists(this.projectFilePath).wait()) {
-				this.$fs.writeJson(this.projectFilePath, {
-					"description": "NativeScript Application",
-					"license": "SEE LICENSE IN <your-license-filename>",
-					"readme": "NativeScript Application",
-					"repository": "<fill-your-repository-here>"
-				}).wait();
-			}
+		if (!this.$fs.exists(this.projectFilePath)) {
+			this.$fs.writeJson(this.projectFilePath, {
+				"description": "NativeScript Application",
+				"license": "SEE LICENSE IN <your-license-filename>",
+				"readme": "NativeScript Application",
+				"repository": "<fill-your-repository-here>"
+			});
+		}
 
-			// Detect indent and use it later to write JSON.
-			let projectFileContent = this.$fs.readText(this.projectFilePath).wait();
+		// Detect indent and use it later to write JSON.
+		let projectFileContent = this.$fs.readText(this.projectFilePath);
 
-			this.projectFileIndent = projectFileContent ? this.detectIndent(projectFileContent) : "\t";
+		this.projectFileIndent = projectFileContent ? this.detectIndent(projectFileContent) : "\t";
 
-			this.projectData = projectFileContent ? JSON.parse(projectFileContent) : Object.create(null);
-
-		}).future<void>()();
+		this.projectData = projectFileContent ? JSON.parse(projectFileContent) : Object.create(null);
 	}
 
 	private detectIndent(content: string): any {
