@@ -6,32 +6,34 @@ temp.track();
 export class ProjectTemplatesService implements IProjectTemplatesService {
 
 	public constructor(private $errors: IErrors,
-						private $fs: IFileSystem,
-						private $logger: ILogger,
-						private $npm: INodePackageManager,
-						private $npmInstallationManager: INpmInstallationManager) { }
+		private $fs: IFileSystem,
+		private $logger: ILogger,
+		private $npm: INodePackageManager,
+		private $npmInstallationManager: INpmInstallationManager,
+		private $analyticsService: IAnalyticsService) { }
 
 	public prepareTemplate(originalTemplateName: string, projectDir: string): IFuture<string> {
 		return ((): string => {
-			let realTemplatePath: string;
-			if(originalTemplateName) {
-				let templateName = originalTemplateName.toLowerCase();
+			let templateName = originalTemplateName || constants.RESERVED_TEMPLATE_NAMES["default"],
+				version: string = null;
 
+			if (originalTemplateName) {
 				// support <reserved_name>@<version> syntax
-				let [name, version] = templateName.split("@");
-				if(constants.RESERVED_TEMPLATE_NAMES[name]) {
-					realTemplatePath = this.prepareNativeScriptTemplate(constants.RESERVED_TEMPLATE_NAMES[name], version, projectDir).wait();
-				} else {
-					// Use the original template name, specified by user as it may be case-sensitive.
-					realTemplatePath = this.prepareNativeScriptTemplate(originalTemplateName, version, projectDir).wait();
-				}
-			} else {
-				realTemplatePath = this.prepareNativeScriptTemplate(constants.RESERVED_TEMPLATE_NAMES["default"], null/*version*/, projectDir).wait();
+				let data = originalTemplateName.split("@"),
+					name = data[0];
+
+				version = data[1];
+
+				templateName = constants.RESERVED_TEMPLATE_NAMES[name.toLowerCase()] || name;
 			}
 
-			if(realTemplatePath) {
+			this.$analyticsService.track("Template used for project creation", templateName).wait();
+
+			const realTemplatePath = this.prepareNativeScriptTemplate(templateName, version, projectDir).wait();
+
+			if (realTemplatePath) {
 				//this removes dependencies from templates so they are not copied to app folder
-				this.$fs.deleteDirectory(path.join(realTemplatePath, constants.NODE_MODULES_FOLDER_NAME)).wait();
+				this.$fs.deleteDirectory(path.join(realTemplatePath, constants.NODE_MODULES_FOLDER_NAME));
 				return realTemplatePath;
 			}
 
@@ -50,7 +52,7 @@ export class ProjectTemplatesService implements IProjectTemplatesService {
 	 */
 	private prepareNativeScriptTemplate(templateName: string, version?: string, projectDir?: string): IFuture<string> {
 		this.$logger.trace(`Using NativeScript verified template: ${templateName} with version ${version}.`);
-		return this.$npmInstallationManager.install(templateName, projectDir, {version: version, dependencyType: "save"});
+		return this.$npmInstallationManager.install(templateName, projectDir, { version: version, dependencyType: "save" });
 	}
 }
 $injector.register("projectTemplatesService", ProjectTemplatesService);
