@@ -18,7 +18,7 @@ class TestExecutionService implements ITestExecutionService {
 		private $projectData: IProjectData,
 		private $platformService: IPlatformService,
 		private $platformsData: IPlatformsData,
-		private $liveSyncProvider: ILiveSyncProvider,
+		private $usbLiveSyncService: ILiveSyncService,
 		private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		private $resources: IResourceLoader,
 		private $httpClient: Server.IHttpClient,
@@ -53,18 +53,19 @@ class TestExecutionService implements ITestExecutionService {
 						this.$options.debugBrk = configOptions.debugBrk;
 						this.$options.debugTransport = configOptions.debugTransport;
 						let configJs = this.generateConfig(this.$options.port.toString(), configOptions);
-						this.$fs.writeFile(path.join(projectDir, TestExecutionService.CONFIG_FILE_NAME), configJs).wait();
+						this.$fs.writeFile(path.join(projectDir, TestExecutionService.CONFIG_FILE_NAME), configJs);
 
 						let socketIoJsUrl = `http://localhost:${this.$options.port}/socket.io/socket.io.js`;
 						let socketIoJs = this.$httpClient.httpRequest(socketIoJsUrl).wait().body;
-						this.$fs.writeFile(path.join(projectDir, TestExecutionService.SOCKETIO_JS_FILE_NAME), socketIoJs).wait();
+						this.$fs.writeFile(path.join(projectDir, TestExecutionService.SOCKETIO_JS_FILE_NAME), socketIoJs);
 
 						if (!this.$platformService.preparePlatform(platform).wait()) {
 							this.$errors.failWithoutHelp("Verify that listed files are well-formed and try again the operation.");
 						}
-						this.detourEntryPoint(projectFilesPath).wait();
+						this.detourEntryPoint(projectFilesPath);
 
-						this.liveSyncProject(platform);
+						this.$platformService.deployPlatform(platform).wait();
+						this.$usbLiveSyncService.liveSync(platform).wait();
 
 						if (this.$options.debugBrk) {
 							this.$logger.info('Starting debugger...');
@@ -103,6 +104,7 @@ class TestExecutionService implements ITestExecutionService {
 
 		let karmaConfig = this.getKarmaConfiguration(platform),
 			karmaRunner = this.$childProcess.fork(path.join(__dirname, "karma-execution.js"));
+
 		karmaRunner.on("message", (karmaData: any) => {
 			fiberBootstrap.run(() => {
 				this.$logger.trace("## Unit-testing: Parent process received message", karmaData);
@@ -111,13 +113,13 @@ class TestExecutionService implements ITestExecutionService {
 					port = karmaData.url.port;
 					let socketIoJsUrl = `http://${karmaData.url.host}/socket.io/socket.io.js`;
 					let socketIoJs = this.$httpClient.httpRequest(socketIoJsUrl).wait().body;
-					this.$fs.writeFile(path.join(projectDir, TestExecutionService.SOCKETIO_JS_FILE_NAME), socketIoJs).wait();
+					this.$fs.writeFile(path.join(projectDir, TestExecutionService.SOCKETIO_JS_FILE_NAME), socketIoJs);
 				}
 
 				if (karmaData.launcherConfig) {
 					let configOptions: IKarmaConfigOptions = JSON.parse(karmaData.launcherConfig);
 					let configJs = this.generateConfig(port, configOptions);
-					this.$fs.writeFile(path.join(projectDir, TestExecutionService.CONFIG_FILE_NAME), configJs).wait();
+					this.$fs.writeFile(path.join(projectDir, TestExecutionService.CONFIG_FILE_NAME), configJs);
 				}
 
 				// Prepare the project AFTER the TestExecutionService.CONFIG_FILE_NAME file is created in node_modules
@@ -129,7 +131,8 @@ class TestExecutionService implements ITestExecutionService {
 				if (this.$options.debugBrk) {
 					this.getDebugService(platform).debug().wait();
 				} else {
-					this.liveSyncProject(platform).wait();
+					this.$platformService.deployPlatform(platform).wait();
+					this.$usbLiveSyncService.liveSync(platform).wait();
 				}
 			});
 		});
@@ -152,13 +155,11 @@ class TestExecutionService implements ITestExecutionService {
 
 	allowedParameters: ICommandParameter[] = [];
 
-	private detourEntryPoint(projectFilesPath: string): IFuture<void> {
-		return (() => {
-			let packageJsonPath = path.join(projectFilesPath, 'package.json');
-			let packageJson = this.$fs.readJson(packageJsonPath).wait();
-			packageJson.main = TestExecutionService.MAIN_APP_NAME;
-			this.$fs.writeJson(packageJsonPath, packageJson).wait();
-		}).future<void>()();
+	private detourEntryPoint(projectFilesPath: string): void {
+		let packageJsonPath = path.join(projectFilesPath, 'package.json');
+		let packageJson = this.$fs.readJson(packageJsonPath);
+		packageJson.main = TestExecutionService.MAIN_APP_NAME;
+		this.$fs.writeJson(packageJsonPath, packageJson);
 	}
 
 	private generateConfig(port: string, options: any): string {
@@ -225,25 +226,6 @@ class TestExecutionService implements ITestExecutionService {
 		this.$logger.debug(JSON.stringify(karmaConfig, null, 4));
 
 		return karmaConfig;
-	}
-
-	private liveSyncProject(platform: string): IFuture<void> {
-		return (() => {
-			let platformData = this.$platformsData.getPlatformData(platform.toLowerCase()),
-				projectFilesPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-
-			let liveSyncData: ILiveSyncData = {
-				platform: platform,
-				appIdentifier: this.$projectData.projectId,
-				projectFilesPath: projectFilesPath,
-				forceExecuteFullSync: true, // Always restart the application when change is detected, so tests will be rerun.
-				syncWorkingDirectory: path.join(this.$projectData.projectDir, constants.APP_FOLDER_NAME),
-				excludedProjectDirsAndFiles: this.$options.release ? constants.LIVESYNC_EXCLUDED_FILE_PATTERNS : []
-			};
-
-			let liveSyncService = this.$injector.resolve(this.$liveSyncProvider.platformSpecificLiveSyncServices[platform.toLowerCase()], { _liveSyncData: liveSyncData });
-			liveSyncService.fullSync().wait();
-		}).future<void>()();
 	}
 }
 $injector.register('testExecutionService', TestExecutionService);
