@@ -34,7 +34,7 @@ describe("Android project properties parser tests", () => {
 
 		let projectPropertiesFileContent = 'target=android-21';
 		let tempFolder = temp.mkdirSync("AndroidProjectPropertiesManager");
-		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent).wait();
+		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent);
 
 		let projectPropertiesManager: IAndroidProjectPropertiesManager = testInjector.resolve(
 			ProjectPropertiesManagerLib.AndroidProjectPropertiesManager, {directoryPath: tempFolder});
@@ -42,7 +42,7 @@ describe("Android project properties parser tests", () => {
 
 		let expectedContent = 'target=android-21' + '\n' +
 		'android.library.reference.1=testValue';
-		let actualContent = fs.readText(path.join(tempFolder, "project.properties")).wait();
+		let actualContent = fs.readText(path.join(tempFolder, "project.properties"));
 
 		assert.equal(expectedContent, actualContent);
 		assert.equal(1, _.keys(projectPropertiesManager.getProjectReferences().wait()).length);
@@ -54,7 +54,7 @@ describe("Android project properties parser tests", () => {
 		let projectPropertiesFileContent = 'target=android-21' + '\n' +
 		'android.library.reference.1=someValue';
 		let tempFolder = temp.mkdirSync("AndroidProjectPropertiesManager");
-		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent).wait();
+		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent);
 
 		let projectPropertiesManager = testInjector.resolve(
 			ProjectPropertiesManagerLib.AndroidProjectPropertiesManager, {directoryPath: tempFolder});
@@ -63,7 +63,7 @@ describe("Android project properties parser tests", () => {
 		let expectedContent = ['target=android-21',
 		'android.library.reference.1=someValue',
 		'android.library.reference.2=testValue'].join('\n');
-		let actualContent = fs.readText(path.join(tempFolder, "project.properties")).wait();
+		let actualContent = fs.readText(path.join(tempFolder, "project.properties"));
 
 		assert.equal(expectedContent, actualContent);
 		assert.equal(2, _.keys(projectPropertiesManager.getProjectReferences().wait()).length);
@@ -79,7 +79,7 @@ describe("Android project properties parser tests", () => {
 		'android.library.reference.4=value4',
 		'android.library.reference.5=value5'].join('\n');
 		let tempFolder = temp.mkdirSync("AndroidProjectPropertiesManager");
-		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent).wait();
+		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent);
 
 		let projectPropertiesManager: IAndroidProjectPropertiesManager = testInjector.resolve(
 			ProjectPropertiesManagerLib.AndroidProjectPropertiesManager, {directoryPath: tempFolder});
@@ -88,7 +88,7 @@ describe("Android project properties parser tests", () => {
 		let expectedContent = projectPropertiesFileContent + '\n' +
 		'android.library.reference.6=testValue';
 
-		let actualContent = fs.readText(path.join(tempFolder, "project.properties")).wait();
+		let actualContent = fs.readText(path.join(tempFolder, "project.properties"));
 
 		assert.equal(expectedContent, actualContent);
 		assert.equal(6, _.keys(projectPropertiesManager.getProjectReferences().wait()).length);
@@ -100,14 +100,14 @@ describe("Android project properties parser tests", () => {
 		let projectPropertiesFileContent = 'android.library.reference.1=value1' + '\n' +
 		'target=android-21';
 		let tempFolder = temp.mkdirSync("AndroidProjectPropertiesManager");
-		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent).wait();
+		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent);
 
 		let projectPropertiesManager: IAndroidProjectPropertiesManager = testInjector.resolve(
 			ProjectPropertiesManagerLib.AndroidProjectPropertiesManager, {directoryPath: tempFolder});
 		projectPropertiesManager.removeProjectReference("value1").wait();
 
 		let expectedContent = 'target=android-21';
-		let actualContent = fs.readText(path.join(tempFolder, "project.properties")).wait();
+		let actualContent = fs.readText(path.join(tempFolder, "project.properties"));
 
 		assert.equal(expectedContent, actualContent);
 		assert.equal(0, _.keys(projectPropertiesManager.getProjectReferences().wait()).length);
@@ -123,7 +123,7 @@ describe("Android project properties parser tests", () => {
 		'android.library.reference.4=value4',
 		'android.library.reference.5=value5'].join('\n');
 		let tempFolder = temp.mkdirSync("AndroidProjectPropertiesManager");
-		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent).wait();
+		fs.writeFile(path.join(tempFolder, "project.properties"), projectPropertiesFileContent);
 
 		let projectPropertiesManager: IAndroidProjectPropertiesManager = testInjector.resolve(
 			ProjectPropertiesManagerLib.AndroidProjectPropertiesManager, {directoryPath: tempFolder});
@@ -134,7 +134,7 @@ describe("Android project properties parser tests", () => {
 		'android.library.reference.2=value2',
 		'android.library.reference.3=value4',
 		'android.library.reference.4=value5'].join('\n') + '\n';
-		let actualContent = fs.readText(path.join(tempFolder, "project.properties")).wait();
+		let actualContent = fs.readText(path.join(tempFolder, "project.properties"));
 
 		assert.equal(expectedContent, actualContent);
 		assert.equal(4, _.keys(projectPropertiesManager.getProjectReferences().wait()).length);
