@@ -1,6 +1,6 @@
-import {Yok} from "../lib/common/yok";
+import { Yok } from "../lib/common/yok";
 import * as stubs from "./stubs";
-import {ProjectTemplatesService} from "../lib/services/project-templates-service";
+import { ProjectTemplatesService } from "../lib/services/project-templates-service";
 import * as assert from "assert";
 import Future = require("fibers/future");
 import * as path from "path";
@@ -9,29 +9,28 @@ import temp = require("temp");
 let isDeleteDirectoryCalledForNodeModulesDir = false;
 let nativeScriptValidatedTemplatePath = "nsValidatedTemplatePath";
 
-function createTestInjector(configuration?: {shouldNpmInstallThrow: boolean, npmInstallationDirContents: string[], npmInstallationDirNodeModulesContents: string[]}): IInjector {
+function createTestInjector(configuration?: { shouldNpmInstallThrow: boolean, npmInstallationDirContents: string[], npmInstallationDirNodeModulesContents: string[] }): IInjector {
 	let injector = new Yok();
 	injector.register("errors", stubs.ErrorsStub);
 	injector.register("logger", stubs.LoggerStub);
 	injector.register("fs", {
-		readDirectory: (dirPath: string) => {
-			if(dirPath.toLowerCase().indexOf("node_modules") !== -1) {
-				return Future.fromResult(configuration.npmInstallationDirNodeModulesContents);
+		readDirectory: (dirPath: string): string[] => {
+			if (dirPath.toLowerCase().indexOf("node_modules") !== -1) {
+				return configuration.npmInstallationDirNodeModulesContents;
 			}
-			return Future.fromResult(configuration.npmInstallationDirContents);
+			return configuration.npmInstallationDirContents;
 		},
 
 		deleteDirectory: (directory: string) => {
-			if(directory.indexOf("node_modules") !== -1) {
+			if (directory.indexOf("node_modules") !== -1) {
 				isDeleteDirectoryCalledForNodeModulesDir = true;
 			}
-			return Future.fromResult();
 		}
 
 	});
 	injector.register("npm", {
 		install: (packageName: string, pathToSave: string, config?: any) => {
-			if(configuration.shouldNpmInstallThrow) {
+			if (configuration.shouldNpmInstallThrow) {
 				throw new Error("NPM install throws error.");
 			}
 
@@ -41,7 +40,7 @@ function createTestInjector(configuration?: {shouldNpmInstallThrow: boolean, npm
 
 	injector.register("npmInstallationManager", {
 		install: (packageName: string, options?: INpmInstallOptions) => {
-			if(configuration.shouldNpmInstallThrow) {
+			if (configuration.shouldNpmInstallThrow) {
 				throw new Error("NPM install throws error.");
 			}
 
@@ -50,6 +49,8 @@ function createTestInjector(configuration?: {shouldNpmInstallThrow: boolean, npm
 	});
 
 	injector.register("projectTemplatesService", ProjectTemplatesService);
+
+	injector.register("analyticsService", { track: () => Future.fromResult() });
 
 	return injector;
 }
@@ -62,9 +63,9 @@ describe("project-templates-service", () => {
 	});
 
 	describe("prepareTemplate", () => {
-		describe("throws error", () =>{
+		describe("throws error", () => {
 			it("when npm install fails", () => {
-				testInjector = createTestInjector({shouldNpmInstallThrow: true, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: null});
+				testInjector = createTestInjector({ shouldNpmInstallThrow: true, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: null });
 				projectTemplatesService = testInjector.resolve("projectTemplatesService");
 				let tempFolder = temp.mkdirSync("preparetemplate");
 				assert.throws(() => projectTemplatesService.prepareTemplate("invalidName", tempFolder).wait());
@@ -72,8 +73,8 @@ describe("project-templates-service", () => {
 		});
 
 		describe("returns correct path to template", () => {
-			it("when reserved template name is used", () =>{
-				testInjector = createTestInjector({shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: []});
+			it("when reserved template name is used", () => {
+				testInjector = createTestInjector({ shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: [] });
 				projectTemplatesService = testInjector.resolve("projectTemplatesService");
 				let tempFolder = temp.mkdirSync("preparetemplate");
 				let actualPathToTemplate = projectTemplatesService.prepareTemplate("typescript", tempFolder).wait();
@@ -81,8 +82,8 @@ describe("project-templates-service", () => {
 				assert.strictEqual(isDeleteDirectoryCalledForNodeModulesDir, true, "When correct path is returned, template's node_modules directory should be deleted.");
 			});
 
-			it("when reserved template name is used (case-insensitive test)", () =>{
-				testInjector = createTestInjector({shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: []});
+			it("when reserved template name is used (case-insensitive test)", () => {
+				testInjector = createTestInjector({ shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: [] });
 				projectTemplatesService = testInjector.resolve("projectTemplatesService");
 				let tempFolder = temp.mkdirSync("preparetemplate");
 				let actualPathToTemplate = projectTemplatesService.prepareTemplate("tYpEsCriPT", tempFolder).wait();
@@ -90,8 +91,8 @@ describe("project-templates-service", () => {
 				assert.strictEqual(isDeleteDirectoryCalledForNodeModulesDir, true, "When correct path is returned, template's node_modules directory should be deleted.");
 			});
 
-			it("uses defaultTemplate when undefined is passed as parameter", () =>{
-				testInjector = createTestInjector({shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: []});
+			it("uses defaultTemplate when undefined is passed as parameter", () => {
+				testInjector = createTestInjector({ shouldNpmInstallThrow: false, npmInstallationDirContents: [], npmInstallationDirNodeModulesContents: [] });
 				projectTemplatesService = testInjector.resolve("projectTemplatesService");
 				let tempFolder = temp.mkdirSync("preparetemplate");
 				let actualPathToTemplate = projectTemplatesService.prepareTemplate(undefined, tempFolder).wait();
